stats

BayesFitness.stats.beta_prior_mutant — Method

beta_prior_mutant(neutrals)

Function to return the vector α̲ for the equivalent of a uniform Dirichlet prior when inferring the relative fitness of a single mutant. Since we use the Beta distribution as the prior when inferring the marginal distribution, this function assigns a 1 to the mutant parameter and a B̲ - 1 to the complement, where B̲ is the total number of unique barcodes.

Arguments

bc_id::Vector{Any}: Vector with the IDs for each barcode

Returns

α̲::Vector{Float64}: Parameters for Beta prior. Mutant lineage is assigned α = 1. The rest of the lineages are grouped together into a single term with αᴮ = B̲ - 1, i.e., the number of lineages minus the mutant lineage being inferred.

BayesFitness.stats.dirichlet_prior_neutral — Method

dirichlet_prior_neutral(neutrals)

Function to return the vector α̲ for the equivalent of a uniform Dirichlet prior when inferring the population mean fitness with the neutral lineages.

Arguments

neutrals::Vector{Bool}: Vector indicating which barcodes correspond to neutral lineages and wich to mutant lineages.

Returns

α̲::Vector{Float64}: Parameters for uniform Dirichlet prior. All lineages lineages are assigned α = 1. The mutant lineages are grouped together into a single term with αᴹ = ∑ₘ α, i.e., the number of non-neutral lineages.

BayesFitness.stats.freq_mutant_ppc — Method

freq_mutant_ppc(df, varname_mut, varname_mean, varname_freq)

Function to compute the posterior predictive checks for the barcode frequency for adaptive mutants. Our model predicts the frequency at time $t+1$ based on the frequency at time $t$ as

\[ f_{t+1}^{(m)} = f_{t}^{(m)} \exp\left[ \left( s^{(m)} - \bar{s}_t \right) \tau \right],\]

where $s^{(m)}$ is the mutant relative fitness, $\bar{s}_t$ is the population mean fitness between time $t$ and $t+1$, and $\tau$ is the time interval between time $t$ and $t+1$. This funciton computes the frequency for each of the MCMC samples in the chain object.

Arguments

df::DataFrames.DataFrame: Dataframe containing the MCMC samples for the variables needed to compute the posterior predictive checks. The dataframe should have MCMC samples for

mutant relative fitness values.

population mean fitness values. NOTE: The number of columns containing population mean fitness values determines the number of datapoints where the ppc are evaluated.

mutant initial frequency.

varname_mut::Union{Symbol, AbstractString}: Variable name for the mutant relative fitness in the data frame.

varname_mean::Union{Symbol, AbstractString}: Variable name pattern for all population mean fitness. All columns in the dataframe should contain this pattern and the sorting of these names must correspond to the sorting of the time points.

varname_freq::Union{Symbol, AbstractString}: Variable name for initial mutant barcode frequencies.

Returns

fₜ₊₁ = fₜ × exp(s⁽ᵐ⁾ - s̅ₜ)::Array{Float64}: Evaluation of the frequency posterior predictive checks at all times for each MCMC sample. The dimensions of the output are (nsamples × ntime × n_chains)

BayesFitness.stats.freq_mutant_ppc — Method

freq_mutant_ppc(chain, varname_mut, varname_mean)

Function to compute the posterior predictive checks for the barcode frequency for adaptive mutants. Our model predicts the frequency at time $t+1$ based on the frequency at time $t$ as

\[ f_{t+1}^{(m)} = f_{t}^{(m)} \exp\left[ \left( s^{(m)} - \bar{s}_t \right) \tau \right],\]

where $s^{(m)}$ is the mutant relative fitness, $\bar{s}_t$ is the population mean fitness between time $t$ and $t+1$, and $\tau$ is the time interval between time $t$ and $t+1$. This funciton computes the frequency for each of the MCMC samples in the chain object.

Arguments

chain::MCMCChains.Chains: Turing.jl MCMC chain for the fitness of a single mutant.

Optional arguments

varname_mut::Symbol=Symbol("s⁽ᵐ⁾"): Variable name for the mutant relative fitness in the chain object.

varname_mean::Symbol=Symbol("s̲ₜ"): Variable name for all population mean fitness.

freq_mut::Symbol=Symbol("f̲⁽ᵐ⁾"): Variable name for all mutant barcode frequencies.

Returns

fₜ₊₁ = fₜ × exp(s⁽ᵐ⁾ - s̅ₜ)::Array{Float64}: Evaluation of the frequency posterior predictive checks at all times for each MCMC sample. The dimensions of the output are (nsamples × ntime × n_chains)

BayesFitness.stats.freq_mutant_ppc_quantile — Method

freq_mut_ppc_quantile(quantile, df, varname_mut, varname_mean, varname_freq)

Function to compute the posterior predictive checks quantiles for the barcode frequency for adaptive mutants. Our model predicts the frequency at time $t+1$ based on the frequency at time $t$ as

\[ f_{t+1}^{(m)} = f_{t}^{(m)} \exp\left[ \left( s^{(m)} - \bar{s}_t \right) \tau \right],\]

where $s^{(m)}$ is the mutant relative fitness, $\bar{s}_t$ is the population mean fitness between time $t$ and $t+1$, and $\tau$ is the time interval between time $t$ and $t+1$. This funciton computes the frequency for each of the MCMC samples in the chain object, and then extracts the quantiles from these posterior predictive checks.

Arguments

quantile::Vector{<:AbstractFloat}: List of quantiles to extract from the posterior predictive checks.

df::DataFrames.DataFrame: Dataframe containing the MCMC samples for the

variables needed to compute the posterior predictive checks. The dataframe should have MCMC samples for

mutant relative fitness values.

population mean fitness values. NOTE: The number of columns containing population mean fitness values determines the number of datapoints where the ppc are evaluated.

mutant initial frequency.

varname_mut::Union{Symbol, AbstractString}: Variable name for the mutant relative fitness in the data frame.

varname_mean::Union{Symbol, AbstractString}: Variable name pattern for all

population mean fitness. All columns in the dataframe should contain this pattern and the sorting of these names must correspond to the sorting of the time points.

varname_freq::Union{Symbol, AbstractString}: Variable name for initial mutant barcode frequencies.

Returns

fₜ₊₁ = fₜ × exp(s⁽ᵐ⁾ - s̅ₜ)::Array{Float64}: Evaluation of the frequency posterior predictive check quantiles at all times for each MCMC sample. The dimensions of the output are (ntime × nquantile × 2), where the last dimension is used to store the lower and upper bound of the quantile. For example, if the quantile is 0.95, the third dimension stores the 0.025 and the 0.975 quantile that encompass the requested 0.95 quantile.

BayesFitness.stats.freq_mutant_ppc_quantile — Method

freq_mut_ppc_quantile(quantile, chain, varname_mut, varname_mean, freq_mut)

Function to compute the posterior predictive checks quantiles for the barcode frequency for adaptive mutants. Our model predicts the frequency at time $t+1$ based on the frequency at time $t$ as

\[ f_{t+1}^{(m)} = f_{t}^{(m)} \exp\left[ \left( s^{(m)} - \bar{s}_t \right) \tau \right],\]

where $s^{(m)}$ is the mutant relative fitness, $\bar{s}_t$ is the population mean fitness between time $t$ and $t+1$, and $\tau$ is the time interval between time $t$ and $t+1$. This funciton computes the frequency for each of the MCMC samples in the chain object, and then extracts the quantiles from these posterior predictive checks.

Arguments

quantile::Vector{<:AbstractFloat}: List of quantiles to extract from the posterior predictive checks.

chain::MCMCChains.Chains: Turing.jl MCMC chain for the fitness of a single mutant.

Optional arguments

varname_mut::Symbol=Symbol("s⁽ᵐ⁾"): Variable name for the mutant relative fitness in the chain object.

varname_mean::Symbol=Symbol("s̲ₜ"): Variable name for all population mean fitness.

freq_mut::Symbol=Symbol("f̲⁽ᵐ⁾"): Variable name for all mutant barcode frequencies.

Returns

fₜ₊₁ = fₜ × exp(s⁽ᵐ⁾ - s̅ₜ)::Array{Float64}: Evaluation of the frequency posterior predictive check quantiles at all times for each MCMC sample. The dimensions of the output are (ntime × nquantile × 2), where the last dimension is used to store the lower and upper bound of the quantile. For example, if the quantile is 0.95, the third dimension stores the 0.025 and the 0.975 quantile that encompass the requested 0.95 quantile.

BayesFitness.stats.gaussian_prior_mean_fitness — Method

gaussian_prior_mean_fitness(data)

Function that fits Gaussian (normal) distributions to MCMC traces from the population mean fitness s̄ₜ. These Gaussians are then used during the mutant relative fitness inference.

Arguments

data::DataFrames.AbstractDataFrame: DataFrame containing the MCMC samples for each of the inferred population mean fitness values, one inferred mean fitness per column.

Returns

µ::Vector{Float64}: Vector encoding the mean values of the Gaussian distributions.

σ::Vector{Float64}: Vector encoding the standard deviation values of the Gaussian distributions.

BayesFitness.stats.logfreqratio_neutral_ppc_quantile — Method

logfreqratio_neutral_ppc_quantile(quantile, df)

Function to compute the posterior predictive checks for the barcode log frequency ratio for neutral lineages. Our model predicts the frequency for neutral lineages at time $t+1$ based on the frequency at time $t$ as

\[ f_{t+1}^{(n)} = f_{t}^{(n)} \exp\left[ - \bar{s}_t \tau \right],\]

where $\bar{s}_t$ is the population mean fitness between time $t$ and $t+1$, and $\tau$ is the time interval between time $t$ and $t+1$. Solving for the mean fitness results in

\[ \frac{1}{\tau} \log \frac{f_{t+1}^{(n)}}{f_{t}^{(n)}} = - \bar{s}_t.\]

This function computes the quantiles of the log frequency ration for the neutral lineages, given the MCMC samples of the population mean fitness.

Arguments

quantile::Vector{<:AbstractFloat}: List of quantiles to extract from the posterior predictive checks.

-df::DataFrames.DataFrame: DataFrame containing all population mean fitness samples⸺multiple chains must be collapsed into a single column⸺one time point per column. Note: we recommend using the var_jld2_to_df from the utils module to build this dataframe.

Returns

log(fₜ₊₁ / fₜ) = - s̅ₜ::Array{Float64}: Evaluation of the log frequency ratio posterior predictive checks at all times for each MCMC sample. The dimensions of the output are (ntime × nquantile × 2), where the last dimension is used to store the lower and upper bound of the quantile. For example, if the quantile is 0.95, the third dimension stores the 0.025 and the 0.975 quantile that encompass the requested 0.95 quantile.