utils

BayesFitness.utils.chain_to_df — Function

chain_to_df(chain, varnames)

Function that extracts the variables in varnames from the chain object and concatenates them into a DataFrame. Multiple chains are piled on top of each other.

Arguments

chain::MCMCChains.Chains: Chain with variables to be converted to dataframe

Optional arguments

varnames::Union{Vector{Symbol}, Nothing}=nothing: Names of the variables in the chain to be extracted. These can be either the full variable name or patterns to extract multi-dimensional variables. If nothing is given, all parameters are extracted into a dataframe.

Returns

DataFrames.DataFrame: Data frame where each column represents all the MCMC samples for a single variable (multiple chains are stack on top of each other)

BayesFitness.utils.var_jld2_concat — Function

var_jld2_concat(dir, pattern, varname)

Function that takes .jld2 files in dir with pattern and extracts a single variable to concatentate into a single chain object. This function is useful to extract variables with the same name on multiple .jld2 files and compile them into a single MCMCChains.Chains object.

NOTE: All chains from which samples will be extracted must have the same number of samples.

Arguments

dir::String: Directory where file(s) with MCMC chains are stored.

pattern::String: Pattern common among all files to process. NOTE: This is use in the Glob.glob command to locate all jld2 files from which to extract the chains.

varname::Symbol: Name of variable in chain object to extract.

Optional Arguments

chainname::String="chain": String defining the dictionary key on the .jld2 file to extract the MCMC chain.

Returns

MCMCChains.Chains: Chain containing all variable samples.

BayesFitness.utils.var_jld2_to_df — Function

var_jld2_to_df(dir, pattern, varname, chainname)

Function that takes .jld2 files in dir with pattern and extracts a single variable into a dataframe. This function is useful to extract, for example, the chains for each inference of the population mean fitness over multiple time points.

NOTE: All chains from which samples will be extracted must have the same number of samples.

Arguments

dir::String: Directory where file(s) with MCMC chains are stored.

pattern::String: Pattern common among all files to process. NOTE: This is use in the Glob.glob command to locate all jld2 files from which to extract the chains.

varname::Symbol: Name of variable in chain object to extract.

Optional Arguments

chainname::String="chain": String defining the dictionary key on the .jld2 file to extract the MCMC chain.

Returns

DataFrames.DataFrame: DataFrame containing all variable samples⸺multiple chains are collapsed into a single column⸺one variable per column.