vi (Variational Inference)

BayesFitness.vi.advi — Method

advi(; kwargs)

Function to sample the joint posterior distribution for the fitness value of all mutant and neutral lineages given a time-series barcode count.

This function expects the data in a tidy format. This means that every row represents a single observation. For example, if we measure barcode i in 4 different time points, each of these four measurements gets an individual row. Furthermore, barcode j measurements over time also get their own individual rows.

The DataFrame must contain at least the following columns:

id_col: Column identifying the ID of the barcode. This can be the barcode sequence, for example.

time_col: Column defining the measurement time point.

count_col: Column with the raw barcode count.

neutral_col: Column indicating whether the barcode is from a neutral lineage

.

Keyword Arguments

data::DataFrames.AbstractDataFrame: Tidy dataframe with the data to be

used to sample from the population mean fitness posterior distribution.

n_walkers::Int: Number of walkers (chains) for the MCMC sample.

n_steps::Int: Number of steps to take.

outputname::String: String to name the .csv output file.

model::Function: Turing.jl model defining the posterior distribution from which to sample (see BayesFitness.model module). This function must take as the first four inputs the following:

R̲̲::Array{Int64}: 2 or 3D array containing the raw barcode counts for all tracked genotypes. The dimensions of this array represent:

dim=1: time.

dim=2: genotype.

dim=3 (optional): experimental repeats

n̲ₜ::VecOrMat{Int64}: Array with the total barcode counts for each time point (on each experimental repeat, if necessary).

n_neutral::Int: Number of neutral lineages.

n_bc::Int: Number of neutral lineages.

Optional Keyword Arguments

model_kwargs::Dict=Dict(): Extra keyword arguments to be passed to the model function.

id_col::Symbol=:barcode: Name of the column in data containing the

barcode identifier. The column may include any type of entry.

time_col::Symbol=:time: Name of the column in data defining the time point at which measurements were done. The column may contain any type of entry as long as sort will result in time-ordered names.

count_col::Symbol=:count: Name of the column in data containing the raw barcode count. The column must contain entries of type Int64.

neutral_col::Symbol=:neutral: Name of the column in data defining whether the barcode belongs to a neutral lineage. The column must contain entries of type Bool.

rep_col::Union{Nothing,Symbol}=nothing: Column indicating the experimental replicate each measurement belongs to. Default is nothing.

env_col::Union{Nothing,Symbol}=nothing: Column indicating the environment in which each measurement was performed. Default is nothing.

genotype_col::Union{Nothing,Symbol}=nothing: Column indicating the genotype each barcode belongs to when fitting a hierarchical model on genotypes. Default is nothing.

rm_T0::Bool=false: Optional argument to remove the first time point from the inference. The data from this first time point is commonly of much lower quality. Therefore, removing this first time point might result in a better inference.

advi::Turing.AdvancedVI.VariationalInference=Tuing.ADVI(1, 10_000): Variational inference algorithm to infer. Currently, Turing.jl only supports ADVI, where the first input is the number of samples to take (empirically one sample works), and the second is the number of update steps to take.

opt::Union{Turing.AdvancedVI.DecayedADAGrad,Flux.Optimise.AbstractOptimiser} = Turing.Variational.DecayedADAGrad(1e-2, 1.1, 0.9): Algorithm used to compute the model gradient and update the parameters. Turing.ADVI can take Flux.jl optimizers. But the recommended algorithm in Stan is the default DecayedADAGrad.

verbose::Bool=true: Boolean indicating if the function should print partial progress to the screen or not.
Returns

df::DataFrames.DataFrame: DataFrame containing summary statistics of

posterior samples for each parameter. Columns include: - mean, std: posterior mean and standard deviation for each variable. - varname: parameter name from the ADVI posterior distribution. - vartype: Description of the type of parameter. The types are: - pop_mean_fitness: Population mean fitness value s̲ₜ. - pop_error: (Nuisance parameter) Log of standard deviation in the likelihood function for the neutral lineages. - bc_fitness: Mutant relative fitness s⁽ᵐ⁾. - bc_hyperfitness: For hierarchical models, mutant hyperparameter that connects the fitness over multiple experimental replicates or multiple genotypes θ⁽ᵐ⁾. - bc_noncenter: (Nuisance parameter) For hierarchical models, non-centered samples used to connect the experimental replicates to the hyperparameter θ̃⁽ᵐ⁾. - bc_deviations: (Nuisance parameter) For hierarchical models, samples that define the log of the deviation from the hyperparameter fitness value logτ⁽ᵐ⁾. - bc_error: (Nuisance parameter) Log of standard deviation in the likelihood function for the mutant lineages. - freq: (Nuisance parameter) Log of the Poisson parameter used to define the frequency of each lineage. - rep: Experimental replicate number. - env: Environment for each parameter. - id: Mutant or neutral strain ID.