viz

BayesFitness.viz.freq_mutant_ppc! — Method

freq_mutant_ppc!(fig, quantile, df, varname_mut, varname_mean, varname_freq; colors, alpha)

Function to plot the posterior predictive checks quantiles for the barcode frequency for adaptive mutants.

Arguments

fig::Makie.Axis: Axis object to be populated with plot.

quantile::Vector{<:AbstractFloat}: List of quantiles to extract from the posterior predictive checks.

chain::MCMCChains.Chains: Turing.jl MCMC chain for the fitness of a single mutant.

df::DataFrames.DataFrame: Dataframe containing the MCMC samples for the

variables needed to compute the posterior predictive checks. The dataframe should have MCMC samples for - mutant relative fitness values. - population mean fitness values. NOTE: The number of columns containing population mean fitness values determines the number of datapoints where the ppc are evaluated. - mutant initial frequency.

varname_mut::Union{Symbol, AbstractString}: Variable name for the mutant relative fitness in the data frame.

varname_mean::Union{Symbol, AbstractString}: Variable name pattern for all population mean fitness. All columns in the dataframe should contain this pattern and the sorting of these names must correspond to the sorting of the time points.

varname_freq::Union{Symbol, AbstractString}: Variable name for initial mutant barcode frequencies.

Optional arguments

colors=ColorSchemes.Blues_9: List of colors to use for each quantile.

alpha::AbstractFloat=0.75: Level of transparency for band representing each quantile.

BayesFitness.viz.freq_mutant_ppc! — Method

freq_mutant_ppc!(fig, quantile, chain; colors, alpha, varname_mut, varname_mean, freq_mut)

Function to plot the posterior predictive checks quantiles for the barcode frequency for adaptive mutants.

Arguments

fig::Makie.Axis: Axis object to be populated with plot.

quantile::Vector{<:AbstractFloat}: List of quantiles to extract from the posterior predictive checks.

chain::MCMCChains.Chains: Turing.jl MCMC chain for the fitness of a single mutant.

Optional arguments

colors=ColorSchemes.Blues_9: List of colors to use for each quantile.

alpha::AbstractFloat=0.75: Level of transparency for band representing each quantile.

varname_mut::Symbol=Symbol("s⁽ᵐ⁾"): Variable name for the mutant relative fitness in the chain object.

varname_mean::Symbol=Symbol("s̲ₜ"): Variable name for all population mean fitness.

freq_mut::Symbol=Symbol("f̲⁽ᵐ⁾"): Variable name for all mutant barcode frequencies.

BayesFitness.viz.logfreqratio_neutral_ppc! — Method

logfreqratio_neutral_ppc!(fig, quantile, df; colors, alpha,)

Function to plot the posterior predictive checks quantiles for the log frequency ratio for neutral lineages

Arguments

fig::Makie.Axis: Axis object to be populated with plot.

quantile::Vector{<:AbstractFloat}: List of quantiles to extract from the posterior predictive checks.

-df::DataFrames.DataFrame: DataFrame containing all population mean fitness samples⸺multiple chains must be collapsed into a single column⸺one time point per column. Note: we recommend using the var_jld2_to_df from the utils module to build this dataframe.

Optional arguments

colors=ColorSchemes.Blues_9: List of colors to use for each quantile.

alpha::AbstractFloat=0.75: Level of transparency for band representing each quantile.

BayesFitness.viz.mcmc_trace_density! — Method

mcmc_trace_density!(fig::Figure, chain::MCMCChains.Chains; colors, labels)

Function to plot the traces and density estimates side-to-side for each of the parametres in the MCMCChains.Chains object.

Arguments

fig::Makie.Figure: Figure object to be populated with plot. This allows the user to decide the size of the figure outside of this function.

chain::MCMCChains.Chains: Samples from the MCMC run generated with Turing.jl.

Optional arguments

colors=ColorSchemes.seaborn_colorblind: List of colors to be used in plot.

labels: List of labels for each of the parameters. If not given, the default will be to use the names stored in the MCMCChains.Chains object.

alpha::AbstractFloat=1: Level of transparency for plots.