model

BayesFitness.model.exprep_fitness_lognormal — Method

exprep_fitness_lognormal(R̲̲, n̲ₜ, n_neutral, n_mut; kwargs)

Turing.jl model to sample the joint posterior distribution for a competitive fitness experiment.

Model

[write model here]

Arguments

R̲̲::Array{Int64, 3}:: T × B × R where T is the number of time points in the data set, B is the number of barcodes, and R is the number of experimental replicates. For each slince in the R axis, each column represents the barcode count trajectory for a single lineage. NOTE: This matrix does not necessarily need to be equivalent to hcat(R̲̲⁽ⁿ⁾, R̲̲⁽ᵐ⁾). This is because R̲̲⁽ᵐ⁾ can exclude mutant barcodes to perform the joint inference only for a subgroup, but R̲̲ must still contain all counts. Usually, if R̲̲⁽ᵐ⁾ excludes mutant barcodes, R̲̲ must be of the form hcat(R̲̲⁽ⁿ⁾, R̲̲⁽ᵐ⁾, R̲̲⁽ᴹ⁾), where R̲̲⁽ᴹ⁾ is a vector that aggregates all excluded mutant barcodes into a "super barcode."

n̲ₜ::Matrix{Int64}: Matrix with the total number of barcode counts for each time point on each replicate. NOTE: This matrix must be equivalent to computing vec(sum(R̲̲, dims=2)). The reason it is an independent input parameter is to avoid the sum computation within the Turing model.

n_neutral::Int: Number of neutral lineages in dataset.

n_mut::Int: Number of mutant lineages in datset.

Optional Keyword Arguments

s_pop_prior::VecOrMat{Float64}=[0.0, 2.0]: Vector or Matrix with the correspnding parameters (Vector: s_pop_prior[1] = mean, s_pop_prior[2] = standard deviation, Matrix: s_pop_prior[:, 1] = mean, s_pop_prior[:, 2] = standard deviation) for a Normal prior on the population mean fitness values. If typeof(s_pop_prior) <: Matrix, there should be as many rows in the matrix as pairs of time adjacent time points in dataset.

σ_pop_prior::VecOrMat{Float64}=[0.0, 1.0]: Vector or Matrix with the correspnding parameters (Vector: σ_pop_prior[1] = mean, σ_pop_prior[2] = standard deviation, Matrix: σ_pop_prior[:, 1] = mean, σ_pop_prior[:, 2] = standard deviation) for a Log-Normal prior on the population mean fitness error utilized in the log-likelihood function. If typeof(σ_pop_prior) <: Matrix, there should be as many rows in the matrix as pairs of time adjacent time points × number of replicates in dataset.

s_mut_prior::VecOrMat{Float64}=[0.0, 2.0]: Vector or Matrix with the correspnding parameters (Vector: s_mut_prior[1] = mean, s_mut_prior[2] = standard deviation, Matrix: s_mut_prior[:, 1] = mean, s_mut_prior[:, 2] = standard deviation) for a Normal prior on the mutant fitness values. If typeof(s_mut_prior) <: Matrix, there should be as many rows in the matrix as number of mutant lineages × number of replicates in the dataset.

σ_mut_prior::VecOrMat{Float64}=[0.0, 1.0]: Vector or Matrix with the correspnding parameters (Vector: s_mut_prior[1] = mean, s_mut_prior[2] = standard deviation, Matrix: s_mut_prior[:, 1] = mean, s_mut_prior[:, 2] = standard deviation) for a Log-Normal prior on the mutant fitness error utilized in the log-likelihood function. If typeof(σ_mut_prior) <: Matrix, there should be as many rows in the matrix as mutant lineages × number of replicates in the dataset.

λ_prior::VecOrMat{Float64}=[3.0, 3.0]: Vector or Matrix with the correspnding parameters (Vector: λ_prior[1] = mean, λ_prior[2] = standard deviation, Matrix: λ_prior[:, 1] = mean, λ_prior[:, 2] = standard deviation) for a Log-Normal prior on the λ parameter in the Poisson distribution. The λ parameter can be interpreted as the mean number of barcode counts since we assume any barcode count n⁽ᵇ⁾ ~ Poisson(λ⁽ᵇ⁾). If typeof(λ_prior) <: Matrix, there should be as many rows in the matrix as number of barcodes × number of time points × number of replicates in the dataset.

BayesFitness.model.exprep_fitness_normal — Method

exprep_fitness_normal(R̲̲, n̲ₜ, n_neutral, n_mut; kwargs)

Turing.jl model to sample the joint posterior distribution for a competitive fitness experiment.

Model

[write model here]

Arguments

R̲̲::Array{Int64, 3}:: T × B × R where T is the number of time points in the data set, B is the number of barcodes, and R is the number of experimental replicates. For each slince in the R axis, each column represents the barcode count trajectory for a single lineage. NOTE: This matrix does not necessarily need to be equivalent to hcat(R̲̲⁽ⁿ⁾, R̲̲⁽ᵐ⁾). This is because R̲̲⁽ᵐ⁾ can exclude mutant barcodes to perform the joint inference only for a subgroup, but R̲̲ must still contain all counts. Usually, if R̲̲⁽ᵐ⁾ excludes mutant barcodes, R̲̲ must be of the form hcat(R̲̲⁽ⁿ⁾, R̲̲⁽ᵐ⁾, R̲̲⁽ᴹ⁾), where R̲̲⁽ᴹ⁾ is a vector that aggregates all excluded mutant barcodes into a "super barcode."

n̲ₜ::Matrix{Int64}: Matrix with the total number of barcode counts for each time point on each replicate. NOTE: This matrix must be equivalent to computing vec(sum(R̲̲, dims=2)). The reason it is an independent input parameter is to avoid the sum computation within the Turing model.

n_neutral::Int: Number of neutral lineages in dataset.

n_mut::Int: Number of mutant lineages in datset.

Optional Keyword Arguments

s_pop_prior::VecOrMat{Float64}=[0.0, 2.0]: Vector or Matrix with the correspnding parameters (Vector: s_pop_prior[1] = mean, s_pop_prior[2] = standard deviation, Matrix: s_pop_prior[:, 1] = mean, s_pop_prior[:, 2] = standard deviation) for a Normal prior on the population mean fitness values. If typeof(s_pop_prior) <: Matrix, there should be as many rows in the matrix as pairs of time adjacent time points in dataset.

logσ_pop_prior::VecOrMat{Float64}=[0.0, 1.0]: Vector or Matrix with the correspnding parameters (Vector: logσ_pop_prior[1] = mean, logσ_pop_prior[2] = standard deviation, Matrix: logσ_pop_prior[:, 1] = mean, logσ_pop_prior[:, 2] = standard deviation) for a Log-Normal prior on the population mean fitness error utilized in the log-likelihood function. If typeof(logσ_pop_prior) <: Matrix, there should be as many rows in the matrix as pairs of time adjacent time points × number of replicates in dataset.

s_mut_prior::VecOrMat{Float64}=[0.0, 2.0]: Vector or Matrix with the correspnding parameters (Vector: s_mut_prior[1] = mean, s_mut_prior[2] = standard deviation, Matrix: s_mut_prior[:, 1] = mean, s_mut_prior[:, 2] = standard deviation) for a Normal prior on the mutant fitness values. If typeof(s_mut_prior) <: Matrix, there should be as many rows in the matrix as number of mutant lineages × number of replicates in the dataset.

logσ_mut_prior::VecOrMat{Float64}=[0.0, 1.0]: Vector or Matrix with the correspnding parameters (Vector: s_mut_prior[1] = mean, s_mut_prior[2] = standard deviation, Matrix: s_mut_prior[:, 1] = mean, s_mut_prior[:, 2] = standard deviation) for a Log-Normal prior on the mutant fitness error utilized in the log-likelihood function. If typeof(logσ_mut_prior) <: Matrix, there should be as many rows in the matrix as mutant lineages × number of replicates in the dataset.

logλ_prior::VecOrMat{Float64}=[3.0, 3.0]: Vector or Matrix with the correspnding parameters (Vector: logλ_prior[1] = mean, logλ_prior[2] = standard deviation, Matrix: logλ_prior[:, 1] = mean, logλ_prior[:, 2] = standard deviation) for a Log-Normal prior on the λ parameter in the Poisson distribution. The λ parameter can be interpreted as the mean number of barcode counts since we assume any barcode count n⁽ᵇ⁾ ~ Poisson(λ⁽ᵇ⁾). If typeof(logλ_prior) <: Matrix, there should be as many rows in the matrix as number of barcodes × number of time points × number of replicates in the dataset.

BayesFitness.model.fitness_lognormal — Method

fitness_lognormal(R̲̲, n̲ₜ, n_neutral, n_mut; kwargs)

Turing.jl model to sample the joint posterior distribution for a competitive fitness experiment.

Model

[write model here]

Arguments

R̲̲::Matrix{Int64}:: T × B matrix–split into a vector of vectors for computational efficiency–where T is the number of time points in the data set and B is the number of barcodes. Each column represents the barcode count trajectory for a single lineage. NOTE: This matrix does not necessarily need to be equivalent to hcat(R̲̲⁽ⁿ⁾, R̲̲⁽ᵐ⁾). This is because R̲̲⁽ᵐ⁾ can exclude mutant barcodes to perform the joint inference only for a subgroup, but R̲̲ must still contain all counts. Usually, if R̲̲⁽ᵐ⁾ excludes mutant barcodes, R̲̲ must be of the form hcat(R̲̲⁽ⁿ⁾, R̲̲⁽ᵐ⁾, R̲̲⁽ᴹ⁾), where R̲̲⁽ᴹ⁾ is a vector that aggregates all excluded mutant barcodes into a "super barcode."

n̲ₜ::Vector{Int64}: Vector with the total number of barcode counts for each time point. NOTE: This vector must be equivalent to computing vec(sum(R̲̲, dims=2)). The reason it is an independent input parameter is to avoid the sum computation within the Turing model.

n_neutral::Int: Number of neutral lineages in dataset.

n_mut::Int: Number of mutant lineages in datset.

Optional Keyword Arguments

s_pop_prior::VecOrMat{Float64}=[0.0, 2.0]: Vector or Matrix with the correspnding parameters (Vector: s_pop_prior[1] = mean, s_pop_prior[2] = standard deviation, Matrix: s_pop_prior[:, 1] = mean, s_pop_prior[:, 2] = standard deviation) for a Normal prior on the population mean fitness values. If typeof(s_pop_prior) <: Matrix, there should be as many rows in the matrix as pairs of time adjacent time points in dataset.

σ_pop_prior::VecOrMat{Float64}=[0.0, 1.0]: Vector or Matrix with the correspnding parameters (Vector: σ_pop_prior[1] = mean, σ_pop_prior[2] = standard deviation, Matrix: σ_pop_prior[:, 1] = mean, σ_pop_prior[:, 2] = standard deviation) for a Log-Normal prior on the population mean fitness error utilized in the log-likelihood function. If typeof(σ_pop_prior) <: Matrix, there should be as many rows in the matrix as pairs of time adjacent time points in dataset.

s_mut_prior::VecOrMat{Float64}=[0.0, 2.0]: Vector or Matrix with the correspnding parameters (Vector: s_mut_prior[1] = mean, s_mut_prior[2] = standard deviation, Matrix: s_mut_prior[:, 1] = mean, s_mut_prior[:, 2] = standard deviation) for a Normal prior on the mutant fitness values. If typeof(s_mut_prior) <: Matrix, there should be as many rows in the matrix as mutant lineages in the dataset.

σ_mut_prior::VecOrMat{Float64}=[0.0, 1.0]: Vector or Matrix with the correspnding parameters (Vector: s_mut_prior[1] = mean, s_mut_prior[2] = standard deviation, Matrix: s_mut_prior[:, 1] = mean, s_mut_prior[:, 2] = standard deviation) for a Log-Normal prior on the mutant fitness error utilized in the log-likelihood function. If typeof(σ_mut_prior) <: Matrix, there should be as many rows in the matrix as mutant lineages in the dataset.

λ_prior::VecOrMat{Float64}=[3.0, 3.0]: Vector or Matrix with the correspnding parameters (Vector: λ_prior[1] = mean, λ_prior[2] = standard deviation, Matrix: λ_prior[:, 1] = mean, λ_prior[:, 2] = standard deviation) for a Log-Normal prior on the λ parameter in the Poisson distribution. The λ parameter can be interpreted as the mean number of barcode counts since we assume any barcode count n⁽ᵇ⁾ ~ Poisson(λ⁽ᵇ⁾). If typeof(λ_prior) <: Matrix, there should be as many rows in the matrix as number of barcodes × number of time points in the dataset.

BayesFitness.model.fitness_normal — Method

fitness_lognormal(R̲̲, n̲ₜ, n_neutral, n_mut; s_pop_prior, logσ_pop_prior, s_mut_prior, logσ_mut_prior, logλ_prior)

Turing.jl model to sample the joint posterior distribution for a competitive fitness experiment.

Model

[write model here]

Arguments

R̲̲::Matrix{Int64}:: T × B matrix–split into a vector of vectors for computational efficiency–where T is the number of time points in the data set and B is the number of barcodes. Each column represents the barcode count trajectory for a single lineage. NOTE: This matrix does not necessarily need to be equivalent to hcat(R̲̲⁽ⁿ⁾, R̲̲⁽ᵐ⁾). This is because R̲̲⁽ᵐ⁾ can exclude mutant barcodes to perform the joint inference only for a subgroup, but R̲̲ must still contain all counts. Usually, if R̲̲⁽ᵐ⁾ excludes mutant barcodes, R̲̲ must be of the form hcat(R̲̲⁽ⁿ⁾, R̲̲⁽ᵐ⁾, R̲̲⁽ᴹ⁾), where R̲̲⁽ᴹ⁾ is a vector that aggregates all excluded mutant barcodes into a "super barcode."

n̲ₜ::Vector{Int64}: Vector with the total number of barcode counts for each time point. NOTE: This vector must be equivalent to computing vec(sum(R̲̲, dims=2)). The reason it is an independent input parameter is to avoid the sum computation within the Turing model.

n_neutral::Int: Number of neutral lineages in dataset.

n_mut::Int: Number of mutant lineages in datset.

Optional Keyword Arguments

s_pop_prior::VecOrMat{Float64}=[0.0, 2.0]: Vector or Matrix with the correspnding parameters (Vector: s_pop_prior[1] = mean, s_pop_prior[2] = standard deviation, Matrix: s_pop_prior[:, 1] = mean, s_pop_prior[:, 2] = standard deviation) for a Normal prior on the population mean fitness values. If typeof(s_pop_prior) <: Matrix, there should be as many rows in the matrix as pairs of time adjacent time points in dataset.

logσ_pop_prior::VecOrMat{Float64}=[0.0, 1.0]: Vector or Matrix with the correspnding parameters (Vector: logσ_pop_prior[1] = mean, logσ_pop_prior[2] = standard deviation, Matrix: logσ_pop_prior[:, 1] = mean, logσ_pop_prior[:, 2] = standard deviation) for a Log-Normal prior on the population mean fitness error utilized in the log-likelihood function. If typeof(logσ_pop_prior) <: Matrix, there should be as many rows in the matrix as pairs of time adjacent time points in dataset.

s_mut_prior::VecOrMat{Float64}=[0.0, 2.0]: Vector or Matrix with the correspnding parameters (Vector: s_mut_prior[1] = mean, s_mut_prior[2] = standard deviation, Matrix: s_mut_prior[:, 1] = mean, s_mut_prior[:, 2] = standard deviation) for a Normal prior on the mutant fitness values. If typeof(s_mut_prior) <: Matrix, there should be as many rows in the matrix as mutant lineages in the dataset.

logσ_mut_prior::VecOrMat{Float64}=[0.0, 1.0]: Vector or Matrix with the correspnding parameters (Vector: s_mut_prior[1] = mean, s_mut_prior[2] = standard deviation, Matrix: s_mut_prior[:, 1] = mean, s_mut_prior[:, 2] = standard deviation) for a Log-Normal prior on the mutant fitness error utilized in the log-likelihood function. If typeof(logσ_mut_prior) <: Matrix, there should be as many rows in the matrix as mutant lineages in the dataset.

logλ_prior::VecOrMat{Float64}=[3.0, 3.0]: Vector or Matrix with the correspnding parameters (Vector: logλ_prior[1] = mean, logλ_prior[2] = standard deviation, Matrix: logλ_prior[:, 1] = mean, logλ_prior[:, 2] = standard deviation) for a Log-Normal prior on the λ parameter in the Poisson distribution. The λ parameter can be interpreted as the mean number of barcode counts since we assume any barcode count n⁽ᵇ⁾ ~ Poisson(λ⁽ᵇ⁾). If typeof(logλ_prior) <: Matrix, there should be as many rows in the matrix as number of barcodes × number of time points in the dataset.

BayesFitness.model.freq_lognormal — Method

freq_lognormal(R̲̲⁽ⁿ⁾, R̲̲⁽ᵐ⁾, R̲̲, n̲ₜ; λ_prior)

Turing.jl model to sample the joint posterior distribution for frequency values on a fitness experiment.

Model

[write model here]

Arguments

R̲̲::Matrix{Int64}:: T × B matrix–split into a vector of vectors for computational efficiency–where T is the number of time points in the data set and B is the number of barcodes. Each column represents the barcode count trajectory for a single lineage. NOTE: This matrix does not necessarily need to be equivalent to hcat(R̲̲⁽ⁿ⁾, R̲̲⁽ᵐ⁾). This is because R̲̲⁽ᵐ⁾ can exclude mutant barcodes to perform the joint inference only for a subgroup, but R̲̲ must still contain all counts. Usually, if R̲̲⁽ᵐ⁾ excludes mutant barcodes, R̲̲ must be of the form hcat(R̲̲⁽ⁿ⁾, R̲̲⁽ᵐ⁾, R̲̲⁽ᴹ⁾), where R̲̲⁽ᴹ⁾ is a vector that aggregates all excluded mutant barcodes into a "super barcode."

n̲ₜ::Vector{Int64}: Vector with the total number of barcode counts for each time point. NOTE: This vector must be equivalent to computing vec(sum(R̲̲, dims=2)). The reason it is an independent input parameter is to avoid the sum computation within the Turing model.

n_neutral::Int: Number of neutral lineages in dataset. NOTE This argument is irrelevant for this function. It is only included to have consistent inputs across models.

n_mut::Int: Number of mutant lineages in datset. NOTE This argument is irrelevant for this function. It is only included to have consistent inputs across models.

Optional Keyword Arguments

λ_prior::Vector{Float64}=[3.0, 3.0]: Vector with the corresponding parameters (λ_prior[1] = mean, λ_prior[2] = standard deviation) for a Log-Normal prior on the λ parameter in the Poisson distribution. The λ parameter can be interpreted as the mean number of barcode counts since we assume any barcode count n⁽ᵇ⁾ ~ Poisson(λ⁽ᵇ⁾). NOTE: This method assigns the same prior to all mutant fitness error values to be inferred.

BayesFitness.model.multienv_fitness_lognormal — Method

multienv_fitness_lognormal(R̲̲, n̲ₜ, n_neutral, n_mut; kwargs)

Turing.jl model to sample the joint posterior distribution for a competitive fitness experiment with different environments on each growth-dilution cycle.

Model

[write model here]

Arguments

R̲̲::Matrix{Int64}:: T × B matrix–split into a vector of vectors for computational efficiency–where T is the number of time points in the data set and B is the number of barcodes. Each column represents the barcode count trajectory for a single lineage. NOTE: This matrix does not necessarily need to be equivalent to hcat(R̲̲⁽ⁿ⁾, R̲̲⁽ᵐ⁾). This is because R̲̲⁽ᵐ⁾ can exclude mutant barcodes to perform the joint inference only for a subgroup, but R̲̲ must still contain all counts. Usually, if R̲̲⁽ᵐ⁾ excludes mutant barcodes, R̲̲ must be of the form hcat(R̲̲⁽ⁿ⁾, R̲̲⁽ᵐ⁾, R̲̲⁽ᴹ⁾), where R̲̲⁽ᴹ⁾ is a vector that aggregates all excluded mutant barcodes into a "super barcode."

n̲ₜ::Vector{Int64}: Vector with the total number of barcode counts for each time point. NOTE: This vector must be equivalent to computing vec(sum(R̲̲, dims=2)). The reason it is an independent input parameter is to avoid the sum computation within the Turing model.

n_neutral::Int: Number of neutral lineages in dataset.

n_mut::Int: Number of mutant lineages in datset.

Keyword Arguments

envs::Vector{<:Any}: List of environments for each time point in dataset. NOTE: The length must be equal to that of n̲ₜ to have one environment per time point.

Optional Keyword Arguments

s_pop_prior::VecOrMat{Float64}=[0.0, 2.0]: Vector or Matrix with the correspnding parameters (Vector: s_pop_prior[1] = mean, s_pop_prior[2] = standard deviation, Matrix: s_pop_prior[:, 1] = mean, s_pop_prior[:, 2] = standard deviation) for a Normal prior on the population mean fitness values. If typeof(s_pop_prior) <: Matrix, there should be as many rows in the matrix as pairs of time adjacent time points in dataset.

σ_pop_prior::VecOrMat{Float64}=[0.0, 1.0]: Vector or Matrix with the correspnding parameters (Vector: σ_pop_prior[1] = mean, σ_pop_prior[2] = standard deviation, Matrix: σ_pop_prior[:, 1] = mean, σ_pop_prior[:, 2] = standard deviation) for a Log-Normal prior on the population mean fitness error utilized in the log-likelihood function. If typeof(σ_pop_prior) <: Matrix, there should be as many rows in the matrix as pairs of time adjacent time points in dataset.

s_mut_prior::VecOrMat{Float64}=[0.0, 2.0]: Vector or Matrix with the correspnding parameters (Vector: s_mut_prior[1] = mean, s_mut_prior[2] = standard deviation, Matrix: s_mut_prior[:, 1] = mean, s_mut_prior[:, 2] = standard deviation) for a Normal prior on the mutant fitness values. If typeof(s_mut_prior) <: Matrix, there should be as many rows in the matrix as mutant lineages × number of unique environments in the dataset.

σ_mut_prior::VecOrMat{Float64}=[0.0, 1.0]: Vector or Matrix with the correspnding parameters (Vector: s_mut_prior[1] = mean, s_mut_prior[2] = standard deviation, Matrix: s_mut_prior[:, 1] = mean, s_mut_prior[:, 2] = standard deviation) for a Log-Normal prior on the mutant fitness error utilized in the log-likelihood function. If typeof(σ_mut_prior) <: Matrix, there should be as many rows in the matrix as mutant lineages × number of unique environments in the dataset.

λ_prior::VecOrMat{Float64}=[3.0, 3.0]: Vector or Matrix with the correspnding parameters (Vector: λ_prior[1] = mean, λ_prior[2] = standard deviation, Matrix: λ_prior[:, 1] = mean, λ_prior[:, 2] = standard deviation) for a Log-Normal prior on the λ parameter in the Poisson distribution. The λ parameter can be interpreted as the mean number of barcode counts since we assume any barcode count n⁽ᵇ⁾ ~ Poisson(λ⁽ᵇ⁾). If typeof(λ_prior) <: Matrix, there should be as many rows in the matrix as number of barcodes × number of time points in the dataset.

BayesFitness.model.neutrals_lognormal — Function

mean_fitness_lognormal(R̲̲, R̲̲⁽ⁿ⁾, n̲ₜ; s_pop_prior, σ_pop_prior, λ_prior)

Turing.jl model to sample the joint posterior distribution of the population mean fitness for a competitive fitness experiment using only the neutral lineages.

Model

[write model here]

Arguments

R̲̲::Matrix{Int64}:: T × B matrix–split into a vector of vectors for computational efficiency–where T is the number of time points in the data set and B is the number of barcodes. Each column represents the barcode count trajectory for a single lineage. NOTE: This matrix does not necessarily need to be equivalent to hcat(R̲̲⁽ⁿ⁾, R̲̲⁽ᵐ⁾). This is because R̲̲⁽ᵐ⁾ can exclude mutant barcodes to perform the joint inference only for a subgroup, but R̲̲ must still contain all counts. Usually, if R̲̲⁽ᵐ⁾ excludes mutant barcodes, R̲̲ must be of the form hcat(R̲̲⁽ⁿ⁾, R̲̲⁽ᵐ⁾, R̲̲⁽ᴹ⁾), where R̲̲⁽ᴹ⁾ is a vector that aggregates all excluded mutant barcodes into a "super barcode."

n̲ₜ::Vector{Int64}: Vector with the total number of barcode counts for each time point. NOTE: This vector must be equivalent to computing vec(sum(R̲̲, dims=2)). The reason it is an independent input parameter is to avoid the sum computation within the Turing model.

n_neutral::Int: Number of neutral lineages in dataset.

n_mut::Int: Number of mutant lineages in datset. NOTE This argument is irrelevant for this function. It is only included to have consistent inputs across models.

Optional Keyword Arguments

s_pop_prior::Vector{Float64}=[0.0, 2.0]: Vector with the correspnding parameters (s_pop_prior[1] = mean, s_pop_prior[2] = standard deviation) for a Normal prior on the population mean fitness values. NOTE: This method assigns the same prior to all population mean fitness to be inferred.

σ_pop_prior::Vector{Float64}=[0.0, 1.0]: Vector with the correspnding parameters (σ_pop_prior[1] = mean, σ_pop_prior[2] = standard deviation) for a Log-Normal prior on the population mean fitness error utilized in the log-likelihood function. NOTE: This method assigns the same prior to all population mean fitness errors to be inferred.

λ_prior::Vector{Float64}=[3.0, 3.0]: Vector with the corresponding parameters (λ_prior[1] = mean, λ_prior[2] = standard deviation) for a Log-Normal prior on the λ parameter in the Poisson distribution. The λ parameter can be interpreted as the mean number of barcode counts since we assume any barcode count n⁽ᵇ⁾ ~ Poisson(λ⁽ᵇ⁾). NOTE: This method assigns the same prior to all mutant fitness error values to be inferred.