model

BayesFitness.model.env_mutant_fitness_lognormal — Method

env_mutant_fitness_lognormal(r̲⁽ᵐ⁾, R̲; α, μ_sₜ, σ_sₜ, s_prior, σ_prior, σ_trunc)

Turing.jl model to sample out of the posterior distribution for a single mutant fitness value s⁽ᵐ⁾, given the raw barcode counts and the parametrization of the population mean fitness distribution.

Arguments

r̲⁽ᵐ⁾::Vector{Int64}: Mutant m raw barcode counts time-series. Note: this vector must be the same length as r̲⁽ᶜ⁾. This means that each entry r̲⁽ᵐ⁾[i] contains the number of reads from barcode m at time i.

R̲::Vector{Int64}: time-series of Raw total reads. This means that entry R̲[i] contains the total number of reads obtained at time i.

Keyword arguments

envs::Vector{<:Any}: Vector defining the order of environments. Environments can be labeled with numbers (e.g. [1, 2, 2, 3, 1, 3]), strings (e.g. ["env1", "env2", "env1"]), or any convenient label. The point being that they should follow the order of environments to which strains were exposed during the experiment.

α::Vector{Float64}: Parameters for Beta prior distribution.

μ_sₜ::Vector{Float64}: Array with the time-series mean values of the population mean fitness. This means entry μ_sₜ[i] contains the inferred mean value of the population mean fitness for time i, assuming sₜ[i] ~ Normal(μ_sₜ[i], σ_sₜ[i]).

σ_sₜ::Vector{Float64}: Array with the time-series values of the population mean fitness standard deviation. This means entry σ_sₜ[i] contains the inferred value of the standard deviation of the population mean fitness at time i, assuming sₜ[i] ~ Normal(μ_sₜ[i], σ_sₜ[i]).

Optional arguments

s_prior::Vector{Real}=[0.0, 2.0]: Parameters for the mutant fitness prior distribution π(s⁽ᵐ⁾).

σ_prior::Vector{Real}=[0.0, 1.0]: Parameters for the nuisance standard deviation parameter prior distribution π(σ⁽ᵐ⁾).

σ_trunc::Real=0.0: Value at which truncate the normal distribution to define it as a half-normal.

BayesFitness.model.fitness_lognormal — Method

fitness_lognormal(R̲̲, R̲̲⁽ⁿ⁾, R̲̲⁽ᵐ⁾, n̲ₜ; s_pop_prior, σ_pop_prior, s_mut_prior, σ_mut_prior, λ_prior)

Turing.jl model to sample the joint posterior distribution for a competitive fitness experiment.

Model

[write model here]

Arguments

R̲̲⁽ⁿ⁾::Matrix{Int64}: T × N matrix where T is the number of time points in the data set and N is the number of neutral lineage barcodes. Each column represents the barcode count trajectory for a single neutral lineage. NOTE: The model assumes the rows are sorted in order of increasing time.

R̲̲⁽ᵐ⁾::Matrix{Int64}: T × M matrix where T is the number of time points in the data set and M is the number of mutant lineage barcodes. Each column represents the barcode count trajectory for a single mutant lineage. NOTE: The model assumes the rows are sorted in order of increasing time.

R̲̲::Matrix{Int64}:: T × B matrix, where T is the number of time points in the data set and B is the number of barcodes. Each column represents the barcode count trajectory for a single lineage. NOTE: This matrix must be equivalent to hcat(R̲̲⁽ⁿ⁾, R̲̲⁽ᵐ⁾). The reason it is an independent input parameter is to avoid the hcat computation within the Turing model.

n̲ₜ::Vector{Int64}: Vector with the total number of barcode counts for each time point. NOTE: This vector must be equivalent to computing vec(sum(R̲̲, dims=2)). The reason it is an independent input parameter is to avoid the sum computation within the Turing model.

Optional Keyword Arguments

s_pop_prior::Vector{Float64}=[0.0, 2.0]: Vector with the correspnding parameters (s_pop_prior[1] = mean, s_pop_prior[2] = standard deviation) for a Normal prior on the population mean fitness values. NOTE: This method assigns the same prior to all population mean fitness to be inferred.

σ_pop_prior::Vector{Float64}=[0.0, 1.0]: Vector with the correspnding parameters (σ_pop_prior[1] = mean, σ_pop_prior[2] = standard deviation) for a Log-Normal prior on the population mean fitness error utilized in the log-likelihood function. NOTE: This method assigns the same prior to all population mean fitness errors to be inferred.

s_mut_prior::Vector{Float64}=[0.0, 2.0]: Vector with the correspnding parameters (s_mut_prior[1] = mean, s_mut_prior[2] = standard deviation) for a Normal prior on the mutant fitness values. NOTE: This method assigns the same prior to all mutant fitness values to be inferred.

σ_mut_prior::Vector{Float64}=[0.0, 1.0]: Vector with the correspnding parameters (σ_mut_prior[1] = mean, σ_mut_prior[2] = standard deviation) for a Log-Normal prior on the mutant fitness error utilized in the log-likelihood function. NOTE: This method assigns the same prior to all mutant fitness error values to be inferred.

λ_prior::Vector{Float64}=[3.0, 3.0]: Vector with the corresponding parameters (λ_prior[1] = mean, λ_prior[2] = standard deviation) for a Log-Normal prior on the λ parameter in the Poisson distribution. The λ parameter can be interpreted as the mean number of barcode counts since we assume any barcode count n⁽ᵇ⁾ ~ Poisson(λ⁽ᵇ⁾). NOTE: This method assigns the same prior to all mutant fitness error values to be inferred.

BayesFitness.model.mean_fitness_neutrals_lognormal — Method

mean_fitness_neutrals_lognormal(r̲ₜ, r̲ₜ₊₁; α, s_prior, σ_prior, σ_trunc)

Turing.jl model to sample the posterior for a single population mean fitness value sₜ, given the raw barcode counts.

Model

For this inference, we can write Bayes theorem as

\[\pi( \bar{s}_t, \sigma_t, \underline{f}_t, \underline{f}_{t+1} \mid \underline{r}_t, \underline{r}_{t+1} ) \propto \prod_{n=1}^N \left[ \pi(f_t^{(n)} \mid \gamma_t^{(n)}) \pi(\gamma_t^{(n)} \mid \bar{s}_t, \sigma_t) \right] \pi(\bar{s}_t) \pi(\sigma_t) \pi(\underline{f}_t \mid \underline{r}_t) \pi(\underline{f}_{t+1} \mid \underline{r}_{t+1})\]

where

\[\gamma_t^{(n)} \equiv \frac{f_{t+1}^{(n)}}{f_t^{n}}.\]

The parametric distributions assumed in this model are of the form

\[f_t^{(n)} \mid \gamma_t^{(n)} \sim \operatorname{Uniform} \left(0, \frac{1}{\gamma_t^{(n)}} \right),\]

\[\gamma_t^{(n)} \mid \bar{s}_t, \sigma_t \sim \log\mathcal{N}(\bar{s}_t, \sigma_t),\]

\[\bar{s}_t \sim \mathcal{N}(\mu_{\bar{s}_t}, \sigma_{\bar{s}_t}),\]

\[\sigma_t \sim \operatorname{Half}-\mathcal{N}(\mu_{\sigma_t}, \sigma_{\sigma_t}),\]

\[\underline{f}_t \mid \underline{r}_t \sim \operatorname{Dirichlet}(\underline{\alpha}_t + \underline{r}_t),\]

and

\[\underline{f}_{t+1} \mid \underline{r}_{t+1} \sim \operatorname{Dirichlet}(\underline{\alpha}_{t+1} + \underline{r}_{t+1}).\]

For this inference, we enforce all frequencies to be > 0 (even for barcodes with zero reads) to compute $\gamma_t^{(n)}$.

The user defines the distribution parameters as:

$\underline{\alpha}_t$: α.

$[\mu_{\bar{s}_t}, \sigma_{\bar{s}_t}]$: s_prior.

$[\mu_{\sigma_t}, \sigma_{\sigma_t}]$: σ_prior.

Arguments

r̲ₜ::Vector{Int64}: Raw counts for neutral lineages and the cumulative counts for mutant lineages at time t. NOTE: The last entry of the array must be the sum of all of the counts from mutant lineages.

r̲ₜ₊₁::Vector{Int64}: Raw counts for neutral lineages and the cumulative counts for mutant lineages at time t + 1. NOTE: The last entry of the array must be the sum of all of the counts from mutant lineages.

Keyword Arguments

α::Vector{Float64}: Parameters for Dirichlet prior distribution.

Optional arguments

s_prior::Vector{Real}=[0.0, 2.0]: Parameters for the mean fitness prior distribution π(sₜ).

σ_prior::Vector{Real}=[0.0, 1.0]: Parameters for the nuisance standard deviation parameter prior distribution π(σₜ).

σ_trunc::Real=0.0: Value at which truncate the normal distribution to define it as a half-normal.

BayesFitness.model.mean_fitness_neutrals_lognormal_priors — Method

mean_fitness_neutrals_lognormal_priors(r̲ₜ, r̲ₜ₊₁; α, s_prior, σ_prior)

Turing.jl model to sample out of the posterior for a single population mean fitness value sₜ, given the raw barcode counts. Note: this function allows for the definition of any prior distributions on the mean fitness and the nuisance standard deviation parameter for the log-likelihood function.

Model

For this inference, we can write Bayes theorem as

\[\pi( \bar{s}_t, \sigma_t, \underline{f}_t, \underline{f}_{t+1} \mid \underline{r}_t, \underline{r}_{t+1} ) \propto \prod_{n=1}^N \left[ \pi(f_t^{(n)} \mid \gamma_t^{(n)}) \pi(\gamma_t^{(n)} \mid \bar{s}_t, \sigma_t) \right] \pi(\bar{s}_t) \pi(\sigma_t) \pi(\underline{f}_t \mid \underline{r}_t) \pi(\underline{f}_{t+1} \mid \underline{r}_{t+1})\]

where

\[\gamma_t^{(n)} \equiv \frac{f_{t+1}^{(n)}}{f_t^{n}}.\]

The parametric distributions assumed in this model are of the form

\[f_t^{(n)} \mid \gamma_t^{(n)} \sim \operatorname{Uniform} \left(0, \frac{1}{\gamma_t^{(n)}} \right),\]

\[\gamma_t^{(n)} \mid \bar{s}_t, \sigma_t \sim \log\mathcal{N}(\bar{s}_t, \sigma_t),\]

\[\bar{s}_t \sim \operatorname{User-defined},\]

\[\sigma_t \sim \operatorname{User-defined},\]

\[\underline{f}_t \mid \underline{r}_t \sim \operatorname{Dirichlet}(\underline{\alpha}_t + \underline{r}_t),\]

and

\[\underline{f}_{t+1} \mid \underline{r}_{t+1} \sim \operatorname{Dirichlet}(\underline{\alpha}_{t+1} + \underline{r}_{t+1}).\]

For this inference, we enforce all frequencies to be > 0 (even for barcodes with zero reads) to compute $\gamma_t^{(n)}$.

The user defines the distribution parameters as:

$\underline{\alpha}_t$: α.

Arguments

r̲ₜ::Vector{Int64}: Raw counts for neutral lineages and the cumulative counts for mutant lineages at time t. NOTE: The last entry of the array must be the sum of all of the counts from mutant lineages.

r̲ₜ₊₁::Vector{Int64}: Raw counts for neutral lineages and the cumulative counts for mutant lineages at time t + 1. NOTE: The last entry of the array must be the sum of all of the counts from mutant lineages.

Keyword Arguments

α::Vector{Float64}: Parameters for Dirichlet prior distribution.

s_prior::Distributions.ContinuousUnivariateDistribution: Parametrized univariate continuous distribution for the prior on the mean fitness π(sₜ).

σ_prior:::Distributions.ContinuousUnivariateDistribution: Parametrized univariate continuous distribution for the prior on the nuisance standard deviation of the log-normal likelihood π(σₜ).

BayesFitness.model.mutant_fitness_lognormal — Method

mutant_fitness_lognormal(r̲⁽ᵐ⁾, R̲; α, μ_sₜ, σ_sₜ, s_prior, σ_prior, σ_trunc)

Turing.jl model to sample out of the posterior distribution for a single mutant fitness value s⁽ᵐ⁾, given the raw barcode counts and the parametrization of the population mean fitness distribution.

Arguments

r̲⁽ᵐ⁾::Vector{Int64}: Mutant m raw barcode counts time-series. Note: this vector must be the same length as r̲⁽ᶜ⁾. This means that each entry r̲⁽ᵐ⁾[i] contains the number of reads from barcode m at time i.

R̲::Vector{Int64}: time-series of Raw total reads. This means that entry R̲[i] contains the total number of reads obtained at time i.

Keyword Arguments

α::Vector{Float64}: Parameters for Beta prior distribution.

μ_sₜ::Vector{Float64}: Array with the time-series mean values of the population mean fitness. This means entry μ_sₜ[i] contains the inferred mean value of the population mean fitness for time i, assuming sₜ[i] ~ Normal(μ_sₜ[i], σ_sₜ[i]).

σ_sₜ::Vector{Float64}: Array with the time-series values of the population mean fitness standard deviation. This means entry σ_sₜ[i] contains the inferred value of the standard deviation of the population mean fitness at time i, assuming sₜ[i] ~ Normal(μ_sₜ[i], σ_sₜ[i]).

Optional arguments

s_prior::Vector{Real}=[0.0, 2.0]: Parameters for the mutant fitness prior distribution π(s⁽ᵐ⁾).

σ_prior::Vector{Real}=[0.0, 1.0]: Parameters for the nuisance standard deviation parameter prior distribution π(σ⁽ᵐ⁾).

σ_trunc::Real=0.0: Value at which truncate the normal distribution to define it as a half-normal.

BayesFitness.model.mutant_fitness_lognormal_priors — Method

mutant_fitness_lognormal_priors(r̲⁽ᵐ⁾, R̲; α, s_mean_priors, s_prior, σ_prior, σ_trunc)

Turing.jl model to sample out of the posterior distribution for a single mutant fitness value s⁽ᵐ⁾, given the raw barcode counts and the parametrization of the population mean fitness distribution. Note: this function allows for the definition of any prior distributions on the population mean fitness, the nuisance standard deviation parameter for the log-likelihood function, and the mutant mean fitness.

Arguments

r̲⁽ᵐ⁾::Vector{Int64}: Mutant m raw barcode counts time-series. Note: this vector must be the same length as r̲⁽ᶜ⁾. This means that each entry r̲⁽ᵐ⁾[i] contains the number of reads from barcode m at time i.

R̲::Vector{Int64}: time-series of Raw total reads. This means that entry R̲[i] contains the total number of reads obtained at time i.

Keyword Arguments

α::Vector{Float64}: Parameters for Beta prior distribution.

s_mean_priors::Vector{<:Distributions.ContinuousUnivariateDistribution}: Vector of univariate distributions defining the prior distribution for each population mean fitness value.

s_prior::Distributions.ContinuousUnivariateDistribution: Parametrized univariate continuous distribution for the prior on the mean fitness π(sₜ).

σ_prior:::Distributions.ContinuousUnivariateDistribution: Parametrized univariate continuous distribution for the prior on the nuisance standard deviation of the log-normal likelihood π(σₜ).