stats

BayesFitness.stats.freq_mutant_ppc — Method

freq_mutant_ppc(n_ppc, df; param, flatten=true)

Function to compute the posterior predictive checks for the barcode frequency for adaptive mutants. Our model predicts the frequency at time $t+1$ based on the frequency at time $t$ as

\[ f_{t+1}^{(m)} = f_{t}^{(m)} \exp\left[ \left( s^{(m)} - \bar{s}_t \right) \tau \right],\]

where $s^{(m)}$ is the mutant relative fitness, $\bar{s}_t$ is the population mean fitness between time $t$ and $t+1$, and $\tau$ is the time interval between time $t$ and $t+1$. Our inference model assumes that

\[ \frac{f_{t+1}^{(m)}}{f_{t}^{(m)}} \sim \log-\mathcal{N}\left( s^{(m)} - \bar{s}_t, \sigma^{(m)} \right),\]

where $\sigma^{(m)}$ is the inferred standard deviation for the model. This function generates samples out of this distribution.

Arguments

df::DataFrames.DataFrame: Dataframe containing the MCMC samples for the variables needed to compute the posterior predictive checks. The dataframe should have MCMC samples for

mutant relative fitness values.

population mean fitness values. NOTE: The number of columns containing population mean fitness values determines the number of datapoints where the ppc are evaluated.

log-normal likelihood standard deviation.

mutant initial frequency.

n_ppc::Int: Number of samples to generate per set of parameters.

Optional Arguments

param::Dict{Symbol, Symbol}: Dictionary indicating the name of the variables in the mcmc chain defining the following variables:

:mutant_mean_fitness: Variable defining the inferred mutant fitness value s⁽ᵐ⁾.

:mutant_std_fitness: Variable defining the standard defining the inferred standard deviation on the likelihood function σ⁽ᵐ⁾.

mutant_freq: Variable defining the inferred initial frequency for the mutant.

population_mean_fitness: Common pattern in all population mean fitness variables.

flatten::Bool=true: Boolean indicating whether to flatten the output of multiple chain into a single column.

Returns

fₜ₊₁ = fₜ × exp(s⁽ᵐ⁾ - s̅ₜ)::Array{Float64}: Evaluation of the frequency posterior predictive checks at all times for each MCMC sample.

BayesFitness.stats.freq_mutant_ppc — Method

freq_mutant_ppc(n_ppc, df; param, flatten=true)

Function to compute the posterior predictive checks for the barcode frequency for adaptive mutants. Our model predicts the frequency at time $t+1$ based on the frequency at time $t$ as

\[ f_{t+1}^{(m)} = f_{t}^{(m)} \exp\left[ \left( s^{(m)} - \bar{s}_t \right) \tau \right],\]

where $s^{(m)}$ is the mutant relative fitness, $\bar{s}_t$ is the population mean fitness between time $t$ and $t+1$, and $\tau$ is the time interval between time $t$ and $t+1$. Our inference model assumes that

\[ \frac{f_{t+1}^{(m)}}{f_{t}^{(m)}} \sim \log-\mathcal{N}\left( s^{(m)} - \bar{s}_t, \sigma^{(m)} \right),\]

where $\sigma^{(m)}$ is the inferred standard deviation for the model. This function generates samples out of this distribution.

Arguments

chain::MCMCChains.Chains: Chain containing the MCMC samples for the variables needed to compute the posterior predictive checks. The dataframe should have MCMC samples for

mutant relative fitness values.

population mean fitness values. NOTE: The number of columns containing population mean fitness values determines the number of datapoints where the ppc are evaluated.

log-normal likelihood standard deviation.

mutant initial frequency.

n_ppc::Int: Number of samples to generate per set of parameters.

Optional Arguments

param::Dict{Symbol, Symbol}: Dictionary indicating the name of the variables in the mcmc chain defining the following variables:

:mutant_mean_fitness: Variable defining the inferred mutant fitness value s⁽ᵐ⁾.

:mutant_std_fitness: Variable defining the standard defining the inferred standard deviation on the likelihood function σ⁽ᵐ⁾.

mutant_freq: Variable defining the inferred initial frequency for the mutant.

population_mean_fitness: Common pattern in all population mean fitness variables.

flatten::Bool=true: Boolean indicating whether to flatten the output of multiple chain into a single column.

Returns

fₜ₊₁ = fₜ × exp(s⁽ᵐ⁾ - s̅ₜ)::Array{Float64}: Evaluation of the frequency posterior predictive checks at all times for each MCMC sample.

BayesFitness.stats.logfreq_ratio_mean_ppc — Method

logfreq_ratio_mean_ppc(df, n_ppc, param; flatten=true)

Function to compute the posterior predictive checks (better called the posterior retrodictive checks) for the barcode log-frequency ratio for neutral lineages. Our model predicts the frequency at time $t+1$ based on the frequency at time $t$ as

\[ f_{t+1}^{(n)} = f_{t}^{(n)} \exp\left[ \left( - \bar{s}_t \right) \tau \right],\]

where $\bar{s}_t$ is the population mean fitness between time $t$ and $t+1$, and $\tau$ is the time interval between time $t$ and $t+1$. Our inference model assumes that

\[ \frac{f_{t+1}^{(n)}}{f_{t}^{(n)}} \sim \log-\mathcal{N}\left( - \bar{s}_t, \sigma^{(n)} \right),\]

where $\sigma^{(n)}$ is the inferred standard deviation for the model. This function generates samples out of this distribution.

Arguments

n_ppc::Int: Number of samples to generate per set of parameters.

df::DataFrames.DataFrame: Dataframe containing the MCMC samples for the variables needed to compute the posterior predictive checks. The dataframe should have MCMC samples for

population mean fitness values. NOTE: The number of columns containing population mean fitness values determines the number of datapoints where the ppc are evaluated.

log-normal likelihood standard deviation.

Optional Arguments

param::Dict{Symbol, Symbol}: Dictionary indicating the name of the variables

in the mcmc chain defining the following variables: - population_mean_fitness: Common pattern in all population mean fitness variables. - population_std_fitness: Common pattern in all standard deviations estimates for the likelihood.

flatten::Bool=true: Boolean indicating whether to flatten the output of multiple chain into a single column.

Returns

log(fₜ₊₁ / fₜ)= s⁽ᵐ⁾ - s̅ₜ::Array{Float64}: Evaluation of the log frequency ratio posterior predictive checks at all times for each MCMC sample.

BayesFitness.stats.logfreq_ratio_mean_ppc — Method

logfreq_ratio_mean_ppc(chain, n_ppc; param, flatten=true)

Function to compute the posterior predictive checks (better called the posterior retrodictive checks) for the barcode log-frequency ratio for neutral lineages. Our model predicts the frequency at time $t+1$ based on the frequency at time $t$ as

\[ f_{t+1}^{(n)} = f_{t}^{(n)} \exp\left[ \left( - \bar{s}_t \right) \tau \right],\]

where $\bar{s}_t$ is the population mean fitness between time $t$ and $t+1$, and $\tau$ is the time interval between time $t$ and $t+1$. Our inference model assumes that

\[ \frac{f_{t+1}^{(n)}}{f_{t}^{(n)}} \sim \log-\mathcal{N}\left( - \bar{s}_t, \sigma^{(n)} \right),\]

where $\sigma^{(n)}$ is the inferred standard deviation for the model. This function generates samples out of this distribution.

Arguments

n_ppc::Int: Number of samples to generate per set of parameters.

df::DataFrames.DataFrame: Dataframe containing the MCMC samples for the variables needed to compute the posterior predictive checks. The dataframe should have MCMC samples for

population mean fitness values. NOTE: The number of columns containing

population mean fitness values determines the number of datapoints where the ppc are evaluated.

log-normal likelihood standard deviation.

Optional Arguments

param::Dict{Symbol, Symbol}: Dictionary indicating the name of the variables

in the mcmc chain defining the following variables: - population_mean_fitness: Common pattern in all population mean fitness variables. - population_std_fitness: Common pattern in all standard deviations estimates for the likelihood.

flatten::Bool=true: Boolean indicating whether to flatten the output of multiple chain into a single column.

Returns

log(fₜ₊₁ / fₜ) = s⁽ᵐ⁾ - s̅ₜ::Array{Float64}: Evaluation of the log frequency ratio posterior predictive checks at all times for each MCMC sample.

BayesFitness.stats.logfreq_ratio_multienv_ppc — Method

logfreq_ratio_mutienv_ppc(df, n_ppc, param; flatten=true)

Function to compute the posterior predictive checks for the barcode log-frequency ratio for adaptive mutants. Our model predicts the frequency at time $t+1$ based on the frequency at time $t$ as

\[ f_{t+1}^{(m)} = f_{t}^{(m)} \exp\left[ \left( s^{(m)} - \bar{s}_t \right) \tau \right],\]

where $s^{(m)}$ is the mutant relative fitness, $\bar{s}_t$ is the population mean fitness between time $t$ and $t+1$, and $\tau$ is the time interval between time $t$ and $t+1$. Our inference model assumes that

\[ \frac{f_{t+1}^{(m)}}{f_{t}^{(m)}} \sim \log-\mathcal{N}\left( s^{(m)} - \bar{s}_t, \sigma^{(m)} \right),\]

where $\sigma^{(m)}$ is the inferred standard deviation for the model. This function generates samples out of this distribution.

Arguments

df::DataFrames.DataFrame: Dataframe containing the MCMC samples for the variables needed to compute the posterior predictive checks. The dataframe should have MCMC samples for

mutant relative fitness values.

population mean fitness values. NOTE: The number of columns containing population mean fitness values determines the number of datapoints where the ppc are evaluated.

log-normal likelihood standard deviation.

mutant initial frequency.

n_ppc::Int: Number of samples to generate per set of parameters.

envs::Vector{<:Any}: List of environments in experiment. This is used to index the corresponding fitness from the chain. NOTE: The list of environments should be the name or corresponding label of the environemnt; the index is generated internally.

Optional Keyword Arguments

param::Dict{Symbol, Symbol}: Dictionary indicating the name of the variables

in the mcmc chain defining the following variables:

:mutant_mean_fitness: Variable defining the inferred mutant fitness value s⁽ᵐ⁾.

:mutant_std_fitness: Variable defining the standard defining the inferred standard deviation on the likelihood function σ⁽ᵐ⁾.

population_mean_fitness: Common pattern in all population mean fitness variables.

flatten::Bool=true: Boolean indicating whether to flatten the output of

multiple chain into a single column.

Returns

log(fₜ₊₁ / fₜ) = s⁽ᵐ⁾ - s̅ₜ::Array{Float64}: Evaluation of the frequency posterior predictive checks at all times for each MCMC sample.

BayesFitness.stats.logfreq_ratio_mutant_ppc — Method

logfreq_ratio_mutant_ppc(df, n_ppc; param, flatten=true)

Function to compute the posterior predictive checks for the barcode log-frequency ratio for adaptive mutants. Our model predicts the frequency at time $t+1$ based on the frequency at time $t$ as

\[ f_{t+1}^{(m)} = f_{t}^{(m)} \exp\left[ \left( s^{(m)} - \bar{s}_t \right) \tau \right],\]

where $s^{(m)}$ is the mutant relative fitness, $\bar{s}_t$ is the population mean fitness between time $t$ and $t+1$, and $\tau$ is the time interval between time $t$ and $t+1$. Our inference model assumes that

\[ \frac{f_{t+1}^{(m)}}{f_{t}^{(m)}} \sim \log-\mathcal{N}\left( s^{(m)} - \bar{s}_t, \sigma^{(m)} \right),\]

where $\sigma^{(m)}$ is the inferred standard deviation for the model. This function generates samples out of this distribution.

Arguments

df::DataFrames.DataFrame: Dataframe containing the MCMC samples for the variables needed to compute the posterior predictive checks. The dataframe should have MCMC samples for

mutant relative fitness values.

population mean fitness values. NOTE: The number of columns containing population mean fitness values determines the number of datapoints where the ppc are evaluated.

log-normal likelihood standard deviation.

mutant initial frequency.

n_ppc::Int: Number of samples to generate per set of parameters.

Optional Keyword Arguments

param::Dict{Symbol, Symbol}: Dictionary indicating the

name of the variables in the mcmc chain defining the following variables:

:mutant_mean_fitness: Variable defining the inferred mutant fitness value s⁽ᵐ⁾.

:mutant_std_fitness: Variable defining the standard defining the inferred standard deviation on the likelihood function σ⁽ᵐ⁾.

population_mean_fitness: Common pattern in all population mean fitness variables.

flatten::Bool=true: Boolean indicating whether to flatten the output of

multiple chain into a single column.

Returns

log(fₜ₊₁ / fₜ) = s⁽ᵐ⁾ - s̅ₜ::Array{Float64}: Evaluation of the frequency posterior predictive checks at all times for each MCMC sample.

BayesFitness.stats.logfreq_ratio_mutant_ppc — Method

logfreq_ratio_mutant_ppc(chain, n_ppc; param, flatten=true)

Function to compute the posterior predictive checks for the barcode log-frequency ratio for adaptive mutants. Our model predicts the frequency at time $t+1$ based on the frequency at time $t$ as

\[ f_{t+1}^{(m)} = f_{t}^{(m)} \exp\left[ \left( s^{(m)} - \bar{s}_t \right) \tau \right],\]

where $s^{(m)}$ is the mutant relative fitness, $\bar{s}_t$ is the population mean fitness between time $t$ and $t+1$, and $\tau$ is the time interval between time $t$ and $t+1$. Our inference model assumes that

\[ \frac{f_{t+1}^{(m)}}{f_{t}^{(m)}} \sim \log-\mathcal{N}\left( s^{(m)} - \bar{s}_t, \sigma^{(m)} \right),\]

where $\sigma^{(m)}$ is the inferred standard deviation for the model. This function generates samples out of this distribution.

Arguments

chain::MCMCChains.Chains: Chain containing the MCMC samples for the variables needed to compute the posterior predictive checks. The dataframe should have MCMC samples for

mutant relative fitness values.

population mean fitness values. NOTE: The number of columns containing population mean fitness values determines the number of datapoints where the ppc are evaluated.

log-normal likelihood standard deviation.

mutant initial frequency.

n_ppc::Int: Number of samples to generate per set of parameters.

Optional Arguments

param::Dict{Symbol, Symbol}: Dictionary indicating the

name of the variables in the mcmc chain defining the following variables:

:mutant_mean_fitness: Variable defining the inferred mutant fitness value s⁽ᵐ⁾.

:mutant_std_fitness: Variable defining the standard defining the inferred standard deviation on the likelihood function σ⁽ᵐ⁾.

population_mean_fitness: Common pattern in all population mean fitness variables.

flatten::Bool=true: Boolean indicating whether to flatten the output of

multiple chain into a single column.

Returns

log(fₜ₊₁ / fₜ) = s⁽ᵐ⁾ - s̅ₜ::Array{Float64}: Evaluation of the frequency posterior predictive checks at all times for each MCMC sample.

BayesFitness.stats.matrix_quantile_range — Method

matrix_quantile_range(quantile, matrix; dim=2)

Function to compute the quantile ranges of matrix mat over dimension dim. For example, if quantile[1] = 0.95, This function returns the 0.025 and 0.975 quantiles that capture 95 percent of the entires on the matrix.

Arguments

quantile::Vector{<:AbstractFloat}: List of quantiles to extract from the posterior predictive checks.

matrix::Matrix{<:Real}: Array over which to compute quantile ranges.

Optional arguments

dim::Int=2: Dimension over which to compute quantiles. Defualt = 1, i.e., columns.

BayesFitness.stats.naive_fitness — Method

naive_fitness(data; id_col, time_col, count_col, neutral_col, pseudo_count)

Function to compute a naive estimate of mutant fitness data based on counts. The fitness estimate is computed as

\[\left\langle \log\frac{f^{(m)}_{t+1}}{f^{(m)}_{t}} - \log\frac{f^{(n)}_{t+1}}{f^{(n)}_{t}} \right\rangle = s^{(m)}\]

Arguments

data::DataFrames.AbstractDataFrame: Tidy dataframe with the data to be

used to infer the fitness values on mutants. The DataFrame must contain at least the following columns: - id_col: Column identifying the ID of the barcode. This can the barcode sequence, for example. - time_col: Column defining the measurement time point. - count_col: Column with the raw barcode count. - neutral_col: Column indicating whether the barcode is from a neutral lineage or not.

Optional Keyword Arguments

id_col::Symbol=:barcode: Name of the column in data containing the barcode identifier. The column may contain any type of entry.

time_col::Symbol=:time: Name of the column in data defining the time point at which measurements were done. The column may contain any type of entry as long as sort will resulted in time-ordered names.

count_col::Symbol=:count: Name of the column in data containing the raw barcode count. The column must contain entries of type Int64.

neutral_col::Symbol=:neutral: Name of the column in data defining whether the barcode belongs to a neutral lineage or not. The column must contain entries of type Bool.

rm_T0::Bool=false: Optional argument to remove the first time point from the

inference. Commonly, the data from this first time point is of much lower quality. Therefore, removing this first time point might result in a better inference.

pseudo_count::Int=1: Pseudo count number to add to all counts. This is useful to avoid divisions by zero.

Returns

DataFrames.DataFrame: Data frame with two columns:

id_col: Column indicating the strain ID.

fitness: Naive fitness estimate.