stats

BayesFitness.stats.beta_prior_mutant — Method

beta_prior_mutant(neutrals)

Function to return the vector α̲ for the equivalent of a uniform Dirichlet prior when inferring the relative fitness of a single mutant. Since we use the Beta distribution as the prior when inferring the marginal distribution, this function assigns a 1 to the mutant parameter and a B̲ - 1 to the complement, where B̲ is the total number of unique barcodes.

Arguments

bc_id::Vector{Any}: Vector with the IDs for each barcode

Returns

α̲::Vector{Float64}: Parameters for Beta prior. Mutant lineage is assigned α = 1. The rest of the lineages are grouped together into a single term with αᴮ = B̲ - 1, i.e., the number of lineages minus the mutant lineage being inferred.

BayesFitness.stats.dirichlet_prior_neutral — Method

dirichlet_prior_neutral(neutrals)

Function to return the vector α̲ for the equivalent of a uniform Dirichlet prior when inferring the population mean fitness with the neutral lineages.

Arguments

neutrals::Vector{Bool}: Vector indicating which barcodes correspond to neutral lineages and wich to mutant lineages.

Returns

α̲::Vector{Float64}: Parameters for uniform Dirichlet prior. All lineages lineages are assigned α = 1. The mutant lineages are grouped together into a single term with αᴹ = ∑ₘ α, i.e., the number of non-neutral lineages.

BayesFitness.stats.freq_mutant_ppc — Method

freq_mutant_ppc(n_ppc, df; param, flatten=true)

Function to compute the posterior predictive checks for the barcode frequency for adaptive mutants. Our model predicts the frequency at time $t+1$ based on the frequency at time $t$ as

\[ f_{t+1}^{(m)} = f_{t}^{(m)} \exp\left[ \left( s^{(m)} - \bar{s}_t \right) \tau \right],\]

where $s^{(m)}$ is the mutant relative fitness, $\bar{s}_t$ is the population mean fitness between time $t$ and $t+1$, and $\tau$ is the time interval between time $t$ and $t+1$. Our inference model assumes that

\[ \frac{f_{t+1}^{(m)}}{f_{t}^{(m)}} \sim \log-\mathcal{N}\left( s^{(m)} - \bar{s}_t, \sigma^{(m)} \right),\]

where $\sigma^{(m)}$ is the inferred standard deviation for the model. This function generates samples out of this distribution.

Arguments

df::DataFrames.DataFrame: Dataframe containing the MCMC samples for the variables needed to compute the posterior predictive checks. The dataframe should have MCMC samples for

mutant relative fitness values.

population mean fitness values. NOTE: The number of columns containing population mean fitness values determines the number of datapoints where the ppc are evaluated.

log-normal likelihood standard deviation.

mutant initial frequency.

n_ppc::Int: Number of samples to generate per set of parameters.

Optional Arguments

param::Dict{Symbol, Symbol}: Dictionary indicating the name of the variables in the mcmc chain defining the following variables:

:mutant_mean_fitness: Variable defining the inferred mutant fitness value s⁽ᵐ⁾.

:mutant_std_fitness: Variable defining the standard defining the inferred standard deviation on the likelihood function σ⁽ᵐ⁾.

mutant_freq: Variable defining the inferred initial frequency for the mutant.

population_mean_fitness: Common pattern in all population mean fitness variables.

flatten::Bool=true: Boolean indicating whether to flatten the output of multiple chain into a single column.

Returns

fₜ₊₁ = fₜ × exp(s⁽ᵐ⁾ - s̅ₜ)::Array{Float64}: Evaluation of the frequency posterior predictive checks at all times for each MCMC sample.

BayesFitness.stats.freq_mutant_ppc — Method

freq_mutant_ppc(n_ppc, df; param, flatten=true)

Function to compute the posterior predictive checks for the barcode frequency for adaptive mutants. Our model predicts the frequency at time $t+1$ based on the frequency at time $t$ as

\[ f_{t+1}^{(m)} = f_{t}^{(m)} \exp\left[ \left( s^{(m)} - \bar{s}_t \right) \tau \right],\]

where $s^{(m)}$ is the mutant relative fitness, $\bar{s}_t$ is the population mean fitness between time $t$ and $t+1$, and $\tau$ is the time interval between time $t$ and $t+1$. Our inference model assumes that

\[ \frac{f_{t+1}^{(m)}}{f_{t}^{(m)}} \sim \log-\mathcal{N}\left( s^{(m)} - \bar{s}_t, \sigma^{(m)} \right),\]

where $\sigma^{(m)}$ is the inferred standard deviation for the model. This function generates samples out of this distribution.

Arguments

chain::MCMCChains.Chains: Chain containing the MCMC samples for the variables needed to compute the posterior predictive checks. The dataframe should have MCMC samples for

mutant relative fitness values.

population mean fitness values. NOTE: The number of columns containing population mean fitness values determines the number of datapoints where the ppc are evaluated.

log-normal likelihood standard deviation.

mutant initial frequency.

n_ppc::Int: Number of samples to generate per set of parameters.

Optional Arguments

param::Dict{Symbol, Symbol}: Dictionary indicating the name of the variables in the mcmc chain defining the following variables:

:mutant_mean_fitness: Variable defining the inferred mutant fitness value s⁽ᵐ⁾.

:mutant_std_fitness: Variable defining the standard defining the inferred standard deviation on the likelihood function σ⁽ᵐ⁾.

mutant_freq: Variable defining the inferred initial frequency for the mutant.

population_mean_fitness: Common pattern in all population mean fitness variables.

flatten::Bool=true: Boolean indicating whether to flatten the output of multiple chain into a single column.

Returns

fₜ₊₁ = fₜ × exp(s⁽ᵐ⁾ - s̅ₜ)::Array{Float64}: Evaluation of the frequency posterior predictive checks at all times for each MCMC sample.

BayesFitness.stats.gaussian_prior_mean_fitness — Method

gaussian_prior_mean_fitness(data)

Function that fits Gaussian (normal) distributions to MCMC traces from the population mean fitness s̄ₜ. These Gaussians are then used during the mutant relative fitness inference.

Arguments

data::DataFrames.AbstractDataFrame: DataFrame containing the MCMC samples for each of the inferred population mean fitness values, one inferred mean fitness per column.

Optional arguments

params::Bool=true: Boolean indicating whether the distribution parameters (mean and variance) or the full distribution should be returned

Returns

if params == true: - µ::Vector{Float64}: Vector encoding the mean values of the Gaussian distributions. - σ::Vector{Float64}: Vector encoding the standard deviation values of the Gaussian distributions. else - dists::Vector{<:Distributions.ContinuousUnivariateDistribution}: Vector with the Distributions.jl fit distributions.

BayesFitness.stats.gaussian_prior_mean_fitness — Method

gaussian_prior_mean_fitness(data)

Function that fits Gaussian (normal) distributions to MCMC traces from the population mean fitness s̄ₜ. These Gaussians are then used during the mutant relative fitness inference.

Arguments

data::DataFrames.AbstractDataFrame: DataFrame containing the MCMC samples for each of the inferred population mean fitness values, one inferred mean fitness per column.

Optional arguments

params::Bool=true: Boolean indicating whether the distribution parameters (mean and variance) or the full distribution should be returned

Returns

if params == true: - µ::Vector{Float64}: Vector encoding the mean values of the Gaussian distributions. - σ::Vector{Float64}: Vector encoding the standard deviation values of the Gaussian distributions. else - dists::Vector{<:Distributions.ContinuousUnivariateDistribution}: Vector with the Distributions.jl fit distributions.

BayesFitness.stats.logfreq_ratio_mean_ppc — Method

logfreq_ratio_mean_ppc(df, n_ppc, param; flatten=true)

Function to compute the posterior predictive checks (better called the posterior retrodictive checks) for the barcode log-frequency ratio for neutral lineages. Our model predicts the frequency at time $t+1$ based on the frequency at time $t$ as

\[ f_{t+1}^{(n)} = f_{t}^{(n)} \exp\left[ \left( - \bar{s}_t \right) \tau \right],\]

where $\bar{s}_t$ is the population mean fitness between time $t$ and $t+1$, and $\tau$ is the time interval between time $t$ and $t+1$. Our inference model assumes that

\[ \frac{f_{t+1}^{(n)}}{f_{t}^{(n)}} \sim \log-\mathcal{N}\left( - \bar{s}_t, \sigma^{(n)} \right),\]

where $\sigma^{(n)}$ is the inferred standard deviation for the model. This function generates samples out of this distribution.

Arguments

n_ppc::Int: Number of samples to generate per set of parameters.

df::DataFrames.DataFrame: Dataframe containing the MCMC samples for the variables needed to compute the posterior predictive checks. The dataframe should have MCMC samples for

population mean fitness values. NOTE: The number of columns containing population mean fitness values determines the number of datapoints where the ppc are evaluated.

log-normal likelihood standard deviation.

Optional Arguments

param::Dict{Symbol, Symbol}: Dictionary indicating the name of the variables

in the mcmc chain defining the following variables: - population_mean_fitness: Common pattern in all population mean fitness variables. - population_std_fitness: Common pattern in all standard deviations estimates for the likelihood.

flatten::Bool=true: Boolean indicating whether to flatten the output of multiple chain into a single column.

Returns

log(fₜ₊₁ / fₜ)= s⁽ᵐ⁾ - s̅ₜ::Array{Float64}: Evaluation of the log frequency ratio posterior predictive checks at all times for each MCMC sample.

BayesFitness.stats.logfreq_ratio_mean_ppc — Method

logfreq_ratio_mean_ppc(chain, n_ppc; param, flatten=true)

Function to compute the posterior predictive checks (better called the posterior retrodictive checks) for the barcode log-frequency ratio for neutral lineages. Our model predicts the frequency at time $t+1$ based on the frequency at time $t$ as

\[ f_{t+1}^{(n)} = f_{t}^{(n)} \exp\left[ \left( - \bar{s}_t \right) \tau \right],\]

where $\bar{s}_t$ is the population mean fitness between time $t$ and $t+1$, and $\tau$ is the time interval between time $t$ and $t+1$. Our inference model assumes that

\[ \frac{f_{t+1}^{(n)}}{f_{t}^{(n)}} \sim \log-\mathcal{N}\left( - \bar{s}_t, \sigma^{(n)} \right),\]

where $\sigma^{(n)}$ is the inferred standard deviation for the model. This function generates samples out of this distribution.

Arguments

n_ppc::Int: Number of samples to generate per set of parameters.

df::DataFrames.DataFrame: Dataframe containing the MCMC samples for the variables needed to compute the posterior predictive checks. The dataframe should have MCMC samples for

population mean fitness values. NOTE: The number of columns containing

population mean fitness values determines the number of datapoints where the ppc are evaluated.

log-normal likelihood standard deviation.

Optional Arguments

param::Dict{Symbol, Symbol}: Dictionary indicating the name of the variables

in the mcmc chain defining the following variables: - population_mean_fitness: Common pattern in all population mean fitness variables. - population_std_fitness: Common pattern in all standard deviations estimates for the likelihood.

flatten::Bool=true: Boolean indicating whether to flatten the output of multiple chain into a single column.

Returns

log(fₜ₊₁ / fₜ) = s⁽ᵐ⁾ - s̅ₜ::Array{Float64}: Evaluation of the log frequency ratio posterior predictive checks at all times for each MCMC sample.

BayesFitness.stats.logfreq_ratio_mutant_ppc — Method

logfreq_ratio_mutant_ppc(df, n_ppc, param; flatten=true)

Function to compute the posterior predictive checks for the barcode log-frequency ratio for adaptive mutants. Our model predicts the frequency at time $t+1$ based on the frequency at time $t$ as

\[ f_{t+1}^{(m)} = f_{t}^{(m)} \exp\left[ \left( s^{(m)} - \bar{s}_t \right) \tau \right],\]

where $s^{(m)}$ is the mutant relative fitness, $\bar{s}_t$ is the population mean fitness between time $t$ and $t+1$, and $\tau$ is the time interval between time $t$ and $t+1$. Our inference model assumes that

\[ \frac{f_{t+1}^{(m)}}{f_{t}^{(m)}} \sim \log-\mathcal{N}\left( s^{(m)} - \bar{s}_t, \sigma^{(m)} \right),\]

where $\sigma^{(m)}$ is the inferred standard deviation for the model. This function generates samples out of this distribution.

Arguments

df::DataFrames.DataFrame: Dataframe containing the MCMC samples for the variables needed to compute the posterior predictive checks. The dataframe should have MCMC samples for

mutant relative fitness values.

population mean fitness values. NOTE: The number of columns containing population mean fitness values determines the number of datapoints where the ppc are evaluated.

log-normal likelihood standard deviation.

mutant initial frequency.

n_ppc::Int: Number of samples to generate per set of parameters.

Optional Arguments

param::Dict{Symbol, Symbol}: Dictionary indicating the

name of the variables in the mcmc chain defining the following variables:

:mutant_mean_fitness: Variable defining the inferred mutant fitness value s⁽ᵐ⁾.

:mutant_std_fitness: Variable defining the standard defining the inferred standard deviation on the likelihood function σ⁽ᵐ⁾.

population_mean_fitness: Common pattern in all population mean fitness variables.

flatten::Bool=true: Boolean indicating whether to flatten the output of

multiple chain into a single column.

Returns

log(fₜ₊₁ / fₜ) = s⁽ᵐ⁾ - s̅ₜ::Array{Float64}: Evaluation of the frequency posterior predictive checks at all times for each MCMC sample.

BayesFitness.stats.logfreq_ratio_mutant_ppc — Method

logfreq_ratio_mutant_ppc(df, n_ppc; param, flatten=true)

Function to compute the posterior predictive checks for the barcode log-frequency ratio for adaptive mutants. Our model predicts the frequency at time $t+1$ based on the frequency at time $t$ as

\[ f_{t+1}^{(m)} = f_{t}^{(m)} \exp\left[ \left( s^{(m)} - \bar{s}_t \right) \tau \right],\]

where $s^{(m)}$ is the mutant relative fitness, $\bar{s}_t$ is the population mean fitness between time $t$ and $t+1$, and $\tau$ is the time interval between time $t$ and $t+1$. Our inference model assumes that

\[ \frac{f_{t+1}^{(m)}}{f_{t}^{(m)}} \sim \log-\mathcal{N}\left( s^{(m)} - \bar{s}_t, \sigma^{(m)} \right),\]

where $\sigma^{(m)}$ is the inferred standard deviation for the model. This function generates samples out of this distribution.

Arguments

chain::MCMCChains.Chains: Chain containing the MCMC samples for the variables needed to compute the posterior predictive checks. The dataframe should have MCMC samples for

mutant relative fitness values.

population mean fitness values. NOTE: The number of columns containing population mean fitness values determines the number of datapoints where the ppc are evaluated.

log-normal likelihood standard deviation.

mutant initial frequency.

n_ppc::Int: Number of samples to generate per set of parameters.

Optional Arguments

param::Dict{Symbol, Symbol}: Dictionary indicating the

name of the variables in the mcmc chain defining the following variables:

:mutant_mean_fitness: Variable defining the inferred mutant fitness value s⁽ᵐ⁾.

:mutant_std_fitness: Variable defining the standard defining the inferred standard deviation on the likelihood function σ⁽ᵐ⁾.

population_mean_fitness: Common pattern in all population mean fitness variables.

flatten::Bool=true: Boolean indicating whether to flatten the output of

multiple chain into a single column.

Returns

log(fₜ₊₁ / fₜ) = s⁽ᵐ⁾ - s̅ₜ::Array{Float64}: Evaluation of the frequency posterior predictive checks at all times for each MCMC sample.

BayesFitness.stats.matrix_quantile_range — Method

matrix_quantile_range(quantile, matrix; dim=2)

Function to compute the quantile ranges of matrix mat over dimension dim. For example, if quantile[1] = 0.95, This function returns the 0.025 and 0.975 quantiles that capture 95 percent of the entires on the matrix.

Arguments

quantile::Vector{<:AbstractFloat}: List of quantiles to extract from the posterior predictive checks.

matrix::Matrix{<:Real}: Array over which to compute quantile ranges.

Optional arguments

dim::Int=2: Dimension over which to compute quantiles. Defualt = 1, i.e., columns.