utils

BayesFitness.utils.advi_to_df — Method

advitodf(dist, vars, mutids; nrep=1, envs=[1], nsamples=10000)

Convert the output of automatic differentiation variational inference (ADVI) to a tidy dataframe.

Arguments

dist::Distributions.Sampleable: The ADVI posterior sampleable distribution object.

vars::Vector{<:Any}: Vector of variable/parameter names from the ADVI run.

mut_ids::Vector{<:Any}: Vector of mutant strain IDs.

Oprtional Keyword Arguments

n_rep::Int=1: Number of experimental replicates. Default is 1.

envs::Vector{<:Any}=[1]: Vector of environment ids for each timepoint. Default is a single environment [1].

genotypes::Vector{<:Any}=[1]: Vector of genotype IDs for each mutant. This is used for hierarchical models on genotypes. Default is a single genotype.

n_samples::Int=10_000: Number of posterior samples to draw used for hierarchical models. Default is 10,000.

Returns

df::DataFrames.DataFrame: DataFrame containing summary statistics of

posterior samples for each parameter. Columns include: - mean, std: posterior mean and standard deviation for each of the variables. - varname: parameter name from the ADVI posterior distribution. - vartype: Description of the type of parameter. The types are: - pop_mean: Population mean fitness value s̲ₜ. - pop_error: (Nuisance parameter) Log of standard deviation in the likelihood function for the neutral lineages. - mut_fitness: Mutant relative fitness s⁽ᵐ⁾. - mut_hyperfitness: For hierarchical models, mutant hyper parameter that connects the fitness over multiple experimental replicates θ⁽ᵐ⁾. - mut_noncenter: (Nuisance parameter) For hierarchical models, non-centered samples used to connect the experimental replicates to the hyperparameter θ̃⁽ᵐ⁾. - mut_deviations: (Nuisance parameter) For hierarchicaal models, samples that define the log of the deviation from the hyper parameter fitness value logτ⁽ᵐ⁾. - mut_error: (Nuisance parameter) Log of standard deviation in the likelihood function for the mutant lineages. - freq: (Nuisance parameter) Log of the Poisson parameter used to define the frequency of each lineage. - rep: Experimental replicate number. - env: Environment for each parameter. - id: Mutant or neutral strain ID.

Notes

Converts multivariate posterior into summarized dataframe format.

Adds metadata like parameter type, replicate, strain ID, etc.

Can handle models with multiple replicates and environments.

Can handle models with hierarchical structure on genotypes.

Useful for post-processing ADVI results for further analysis and plotting.

BayesFitness.utils.data_to_arrays — Method

data_to_arrays(data; kwargs)

Function to preprocess the tidy dataframe data into the corresponding inputs for the models in the model submodule.

Arguments

data::DataFrames.AbstractDataFrame: Tidy dataframe with the data to be used for sampling the model posterior distribution.

Optional Keyword Arguments

id_col::Symbol=:barcode: Name of the column in data containing the barcode identifier. The column may contain any type of entry.

time_col::Symbol=:time: Name of the column in data defining the time point at which measurements were done. The column may contain any type of entry as long as sort will resulted in time-ordered names.

count_col::Symbol=:count: Name of the column in data containing the raw barcode count. The column must contain entries of type Int64.

neutral_col::Symbol=:neutral: Name of the column in data defining whether the barcode belongs to a neutral lineage or not. The column must contain entries of type Bool.

rm_T0::Bool=false: Optional argument to remove the first time point from the inference. Commonly, the data from this first time point is of much lower quality. Therefore, removing this first time point might result in a better inference.

verbose::Bool=true: Boolean indicating if printing statements should be made.

Returns

if typeof(rep_col) <: Nothing Dictionary with the following entries: - bc_count::Matrix: T × B matrix with all barcodes read counts. - bc_total::Vector: T-dimensional vector with the total number of reads per time point. - n_neutral::Int: Number of neutral lineages. - n_mut::Int: Number of mutant lineages. - mut_keys: List of mutant names in the order used to build R̲̲.

elseif typeof(rep_col) <: Symbol Dictionary with the following entries: - bc_count::Array: T × B × R array with all barcodes read counts. - bc_total::Matrix: T × R matrix with the total number of reads per time point per repeat. - n_neutral::Int: Number of neutral lineages. - n_mut::Int: Number of mutant lineages. - mut_keys: List of mutant names in the order used to build R̲̲.