viz

BayesFitness.viz.bc_time_series! — Method

bc_time_series!(ax, data; color, alpha, id_col, time_col, quant_col)

Function to plot the time series of a quantity (frequency or raw counts, for example) for a set of barcodes. This function expects the data in a tidy format. This means that every row represents a single observation. For example, if we measure barcode i in 4 different time points, each of these four measurements gets an individual row. Furthermore, measurements of barcode j over time also get their own individual rows.

The DataFrame must contain at least the following columns:

id_col: Column identifying the ID of the barcode. This can the barcode sequence, for example.

time_col: Column defining the measurement time point.

quant_col: Column with the quantity to be plot over time.

Arguments

ax::Makie.Axis: Axis object to be populated with plot.

data::DataFrames.AbstractDataFrame: Tidy dataframe with the data to be used to sample from the population mean fitness posterior distribution.

Optional Arguments

id_col::Symbol=:barcode: Name of the column in data containing the barcode identifier. The column may contain any type of entry.

time_col::Symbol=:time: Name of the column in data defining the time point

at which measurements were done. The column may contain any type of entry as long as sort will resulted in time-ordered names.

quant_col::Symbol=:count: Name of the column in data containing the raw barcode count.

zero_lim::Real=1E-8: Number defining under which value quant_col should be considered as zero. These plots are mostly displayed in log scale, thus having a minimum threshold helps with undetermined values.

zero_label::Union{String, Nothing}: Label to be added to the detection limit. If nothing, nothing is added to the plot.

n_ticks::Int: Ideal number of ticks to add to plot. See Makie.WilkinsonTicks.

color::Union{ColorSchemes.ColorScheme,Symbol,ColorTypes.Colorant{Float64, 3}}=ColorSchemes.glasbey_hv_n256: Single color or list of colors from ColorSchemes.jl to be used in plot. Note: when a color list is provided, colors are randomnly assigned to each barcode by sampling from the list of colors.

alpha::AbstractFloat=1.0: Level of transparency for plots.

linewidth::Real=5: Trajectory linewidth.

BayesFitness.viz.logfreq_ratio_time_series! — Method

logfreq_ratio_time_series!((ax, data; color, alpha, id_col, time_col, freq_col)

Function to plot the time series of the log frequency ratiofor a set of barcodes. This function expects the data in a tidy format. This means that every row represents a single observation. For example, if we measure barcode i in 4 different time points, each of these four measurements gets an individual row. Furthermore, measurements of barcode j over time also get their own individual rows.

The DataFrame must contain at least the following columns:

id_col: Column identifying the ID of the barcode. This can the barcode sequence, for example.

time_col: Column defining the measurement time point.

freq_col: Column with the frequency from which to compute the log ratio.

Arguments

ax::Makie.Axis: Axis object to be populated with plot.

data::DataFrames.AbstractDataFrame: Tidy dataframe with the data to be used to sample from the population mean fitness posterior distribution.

Optional Arguments

id_col::Symbol=:barcode: Name of the column in data containing the barcode identifier. The column may contain any type of entry.

time_col::Symbol=:time: Name of the column in data defining the time point

at which measurements were done. The column may contain any type of entry as long as sort will resulted in time-ordered names.

freq_col::Symbol=:count: Name of the column in data containing the barcode frequency.

`color::Union{ColorSchemes.ColorScheme,Symbol,ColorTypes.Colorant{Float64,

3}}=ColorSchemes.glasbeyhvn256: Single color or list of colors fromColorSchemes.jl` to be used in plot. Note: when a color list is provided, colors are randomnly assigned to each barcode by sampling from the list of colors.

alpha::AbstractFloat=1.0: Level of transparency for plots.

linewidth::Real=5: Trajectory linewidth.

log_fn::Union{typeof(log), typeof(log10), typeof(log2)}=log: Log function to be used in plot.

BayesFitness.viz.mcmc_fitdist_cdf! — Method

mcmc_fitdist_cdf!(ax, chain, dist; n_points, range, ecdf_label, cdf_label, ecdf_kwargs, cdf_kwargs, legend, legend_kwargs)

Function to plot the ECDF produced from an MCMC chain along with a fitn distribution. This plot serves to compare if the parametrized distribution matches the density of MCMC samples.

NOTE: For this function ecdf refers to the empirical cumulative distribution function build from the MCMC chain, and cdf refers to the parametric cumulative distribution function from the fit distribution.

Arguments

ax::Makie.Axis: Axis object to be populated with plot.

chain::Vectors{<:Real}: Vector with the MCMC samples from which to build the ECDF plot.

dist::Distributions.ContinuousUnivariateDistribution: Parametric distribution to be compared with the ECDF.

Optional Arguments

npoints::Int=1000: Number of points to evaluate the parametric CDF.

range::Union{Nothing,NTuple{2,<:Real}}=nothing: Range on which to evaluate the parametric CDF. If nothing is provided (default), the range is inferred from the range of MCMC samples.

ecdf_label::String="mcmc": Legend label for the ECDF plot.

cdf_label::String="fit": Legend label for the CDF plot.

ecdf_kwargs::Dict: Extra keyword arguments for the Makie.ecdfplot! function.

cdf_kwargs::Dict: Extra keyword arguments for the Makie.lines! function.

legend::Bool=true: Boolean indicating if a legend should be added to the plot.

legend_kwargs:Dict(): Extra keyword arguments for the Makie.axislegend function.

BayesFitness.viz.mcmc_trace_density! — Method

mcmc_trace_density!(gl::GridLayout, chain::MCMCChains.Chains; colors, labels)

Function to plot the traces and density estimates side-to-side for each of the parametres in the MCMCChains.Chains object.

Arguments

gl::Makie.GridLayout: GridLayout object to be populated with plot. This allows the user to have more flexibility on whether they want to embed this plot within other subplots.

chain::MCMCChains.Chains: Samples from the MCMC run generated with Turing.jl.

Optional arguments

colors=ColorSchemes.seaborn_colorblind: List of colors to be used in plot.

labels: List of labels for each of the parameters. If not given, the default will be to use the names stored in the MCMCChains.Chains object.

alpha::AbstractFloat=1: Level of transparency for plots.

title::Union{String,Nothing}=nothing: Plot title.

title_valign::Symbol=:bottom: Vertical alignment for title label,

title_font::Symbol=:bold: Type of font to be used in plot.

title_fontsize::Real=20: Font size for title.

title_padding::NTuple{4,<:Real}=(0, 0, 5, 0): Padding for plot text.

BayesFitness.viz.mcmc_trace_density! — Method

mcmc_trace_density!(fig::Figure, chain::MCMCChains.Chains; colors, labels)

Function to plot the traces and density estimates side-to-side for each of the parametres in the MCMCChains.Chains object.

Arguments

fig::Makie.Figure: Figure object to be populated with plot. This allows the user to decide the size of the figure outside of this function.

chain::MCMCChains.Chains: Samples from the MCMC run generated with Turing.jl.

Optional arguments

colors=ColorSchemes.seaborn_colorblind: List of colors to be used in plot.

labels: List of labels for each of the parameters. If not given, the default will be to use the names stored in the MCMCChains.Chains object.

alpha::AbstractFloat=1: Level of transparency for plots.

title::Union{String,Nothing}=nothing: Plot title.

title_valign::Symbol=:bottom: Vertical alignment for title label,

title_font::Symbol=:bold: Type of font to be used in plot.

title_fontsize::Real=20: Font size for title.

title_padding::Vector{<:Real}=(0, 0, 5, 0): Padding for plot text.

BayesFitness.viz.ppc_time_series! — Method

ppc_time_series!(ax, quantile, ppc_mat; time, colors, alpha)

Function to plot the posterior predictive checks quantiles for any quantity.

Arguments

ax::Makie.Axis: Axis object to be populated with plot.

quantile::Vector{<:AbstractFloat}: List of quantiles to extract from the posterior predictive checks.

ppc_mat::Matrix{<:AbstractFloat}: Matrix containing the posterior predictive samples. Rows are assumed to contain the samples, columns the time points.

Optional arguments

colors=ColorSchemes.Blues_9: List of colors to use for each quantile.

alpha::AbstractFloat=0.75: Level of transparency for band representing each

quantile.